% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_add.R
\name{add}
\alias{add}
\title{Addition}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number to be added to the first number.}
}
\value{
The sum of the two numbers.
}
\description{
This is a function that performs an addition between two numbers.
}
\examples{

add(1,2)

}
