% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.r
\name{add_2}
\alias{add_2}
\title{add two}
\usage{
add_2(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a numeric
}
\description{
this functions just adds two
}
\details{
there are no specific detail
}
\author{
Daniel Knitter
}
