% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.r
\name{add_3}
\alias{add_3}
\title{add three}
\usage{
add_3(y)
}
\arguments{
\item{y}{a numeric}
}
\value{
a numeric
}
\description{
this functions just adds three
}
\details{
there are no specific details
}
\author{
Daniel Knitter
}
