% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_of_square.R
\name{area_of_square}
\alias{area_of_square}
\title{Rectangle area}
\usage{
area_of_square(length, width)
}
\arguments{
\item{length}{the length of the rectangle}

\item{width}{width of the rectangle}
}
\description{
Function which calculates the area of a square
}
\details{
x
}
\author{
Jessica Loriaux
}
