% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{BenchmarkMetrics}
\alias{BenchmarkMetrics}
\title{Constructor for the BenchmarkMetrics object}
\usage{
BenchmarkMetrics(
  Algorithm = character(),
  Raw_data,
  Metadata,
  Adj_data = list(),
  PCs = list(),
  UMAPs = list(),
  Latent_dims = list(),
  RunningTime = list(),
  Silhouette = list(),
  ARI = list(),
  LISI = list()
)
}
\description{
Constructor for the BenchmarkMetrics object
}
