% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{beta_hat}
\alias{beta_hat}
\title{Estimate beta using the normal equation}
\usage{
beta_hat(X, y)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Outcome vector}
}
\value{
Estimated coefficient vector
}
\description{
Estimate beta using the normal equation
}
\examples{
X <- matrix(c(1, 2, 3, 4), ncol = 2)
y <- c(1, 2)
beta_hat(X, y)
}
