% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcHSI_cat.R
\name{CalcHSI_cat}
\alias{CalcHSI_cat}
\title{Calculate Habitat Suitability Index Table for Gag grouper}
\usage{
CalcHSI_cat(dat, EnvVariable)
}
\arguments{
\item{dat}{Data frame of all catch (positive and zero). This data frame should include columns for Reference, Tot_Num_Gag, and categorical environmental variables (like bottom type)}

\item{EnvVariable}{Categorical environmental variable name (not in quotes)}
}
\value{
returns a table of standardized S (suitability) values for each level of the variable
}
\description{
Calculates HSI for individual categorical environmental variables in the catch data set
}
\examples{
data(example_HSI_dat)
BottomType_HSI <- CalcHSI_cat(dat = example_HSI_dat, EnvVariable = Bottom)
}
