% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUErfa}
\alias{calcmCPUErfa}
\title{calcmCPUErfa}
\usage{
calcmCPUErfa(RFA, species, year, quarter, data, ALK = NULL)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{data}{the datras-data needed.}

\item{ALK}{the ALK.}
}
\value{
Returns the mCPUE per length class in the given roundfish area. TODO: include the ALK to return per age class given ALK instead
}
\description{
Calculates CPUE per length class in a given roundfish area.
}
\examples{

}
