% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUErfaWithALKNew}
\alias{calcmCPUErfaWithALKNew}
\title{calcmCPUErfaWithALKNew}
\usage{
calcmCPUErfaWithALKNew(RFA, species, year, quarter, data, ALKNew,
  procedure = "", percentOfAreaRepresentative = NULL)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{data}{the datras-data needed.}

\item{ALK}{A list with the ALKs. Id to trawl haul is given in the first coumn of each element in the list.}
}
\description{
Calculates CPUE per length class in a given roundfish area with  a new procedure TODO:document.
}
\examples{

}
