% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUEstatRecWithALK}
\alias{calcmCPUEstatRecWithALK}
\title{calcmCPUEstatRecWithALK}
\usage{
calcmCPUEstatRecWithALK(statRec, species, year, quarter, data, ALK,
  percentOfAreaRepresentative = NULL)
}
\arguments{
\item{statRec}{Statistical area.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{data}{the datras-data needed.}

\item{ALK}{the ALK.}

\item{percentOfAreaRepresentative}{the percentage of the statical recangle within sea depth intervall}
}
\value{
Returns the mCPUE per length class in the given statistical rectangle
}
\description{
Calculates CPUE per length class in a given statistical rectangle.
}
