% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUEstatRecWithALKNew}
\alias{calcmCPUEstatRecWithALKNew}
\title{calcmCPUEstatRecWithALKNew}
\usage{
calcmCPUEstatRecWithALKNew(statRec, species, year, quarter, data, ALKNew,
  procedure = "", percentOfAreaRepresentative = NULL)
}
\arguments{
\item{statRec}{Statistical area.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{data}{the datras-data needed.}

\item{percentOfAreaRepresentative}{the percentage of the statical recangle within sea depth intervall}

\item{ALK}{List with the ALKs for each trawl haul. First element of each ALK in the list is a vector with the first element eual the haulId. This haulId is used when finding which of the ALK which are used.}
}
\value{
Returns the mCPUE per length class in the given statistical rectangle
}
\description{
Calculates CPUE per length class in a given statistical rectangle with the new procedure. TODO: document.
}
