% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALK}
\alias{calculateALK}
\title{calculateALK}
\usage{
calculateALK(RFA, species, year, quarter, ca_hh)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALK is calculated.}

\item{quarter}{The quarter of the year which the ALK is calculated.}

\item{ca_hh}{Merged CA and HH, which consist the data needed for calculating the ALK.}
}
\value{
Returns a matrix with the ALK
}
\description{
Calculates the ALK-key-matrix for a given year, quarter, species and RFA.
}
\examples{
calculateALK(RFA = 2,species = "Gadus morhua", year = 2016, quarter = 1)
}
