% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALKDatras}
\alias{calculateALKDatras}
\title{calculateALK}
\usage{
calculateALKDatras(RFA, species, year, quarter, data)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALK is calculated.}

\item{quarter}{The quarter of the year which the ALK is calculated.}

\item{data}{The data needed for calculating the ALK.}
}
\value{
Returns a matrix with the ALK for the given data, species, time and RFA
}
\description{
Calculates the ALK-key-matrix for a given year, quarter, species and RFA.
}
