% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALKHaulbased}
\alias{calculateALKHaulbased}
\title{calculateALKNew}
\usage{
calculateALKHaulbased(RFA, species, year, quarter, data, data_hl,
  dfLength = 1)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALKs are calculated.}

\item{quarter}{The quarter of the year which the ALKs are calculated.}

\item{data}{The CA needed for calculating the ALKs.}

\item{data_hl}{The HL needed for calculating the ALKs (since there can be trawl hauls without age information).}

\item{dfLength}{The length of the pooled length class. Default is 1, e.g. 1 length classes in each pooled length class.}
}
\value{
Returns a list with ALK for each trawl haul
}
\description{
calculateALKNew
}
