% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALKModel}
\alias{calculateALKModel}
\title{calculateALKModel}
\usage{
calculateALKModel(RFA, species, year, quarter, hh, data, fitModel = NULL,
  report = NULL)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALKs are calculated.}

\item{quarter}{The quarter of the year which the ALKs are calculated.}
}
\value{
Returns a list with ALK for each trawl haul
}
\description{
calculateALKModel
}
