% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALKNew}
\alias{calculateALKNew}
\title{calculateALKNew}
\usage{
<<<<<<< HEAD
calculateALKNew(RFA, species, year, quarter, data, data_hl, dfLength = 5)
=======
<<<<<<< HEAD
calculateALKNew(RFA, species, year, quarter, data, data_hl, dfLength = 1)
=======
calculateALKNew(RFA, species, year, quarter, data, data_hl, dfLength = 5)
>>>>>>> f3150d2c6b2352ceaae3b42a8c1ad527a367ca09
>>>>>>> 8dac1c420ad5a870f0d45d364235b2899834e81e
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALKs are calculated.}

\item{quarter}{The quarter of the year which the ALKs are calculated.}

\item{data}{The CA needed for calculating the ALKs.}

\item{data_hl}{The HL needed for calculating the ALKs (since there can be trawl hauls without age information).}

\item{dfLength}{The length of the pooled length class. Default is 5, e.g. 5 length classes in each pooled length class.}
}
\value{
Returns a list with ALK for each trawl haul
}
