% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chk}
\alias{chk}
\title{Standard R cars data set}
\format{
\subsection{'chk'}{

A data frame with 32 rows and 11 variables:
\describe{
\item{mpg}{miles per gallon}
\item{cyl}{number of cylinders}
\item{disp}{engine displacement}
\item{hp}{engine horse power}
\item{drat}{drag racing power}
\item{wt}{weight of car}
\item{qsec}{quality per second}
\item{vs}{vital sign of car}
\item{am}{type of radio}
\item{gear}{number of gears}
\item{carb}{number of carburetors}
}
}
}
\usage{
chk
}
\description{
This is the standard mtcars data set included with base R.
}
\keyword{datasets}
