% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison Analysis.R
\name{comparisonAnalysis}
\alias{comparisonAnalysis}
\title{This function does comparison analysis based on ATAC-seq data.
It assumes that the independent variable is the stimulation.
There are two conditions (GF and SPF).
Four tsv report files will be produced for the following analysis:
Condition GF upregulation
Condition GF downregulation
Condition SPF upregulation
Condition SPF downregulation}
\usage{
comparisonAnalysis(metadataFileName, countsFileName, reference)
}
\arguments{
\item{metadataFileName}{Path to metadata file for ATAC-seq}

\item{countsFileName}{Path to file with read counts}

\item{reference}{Stimulation (Cytokine or PBS) that will be used as the reference}
}
\value{
None
}
\description{
This function does comparison analysis based on ATAC-seq data.
It assumes that the independent variable is the stimulation.
There are two conditions (GF and SPF).
Four tsv report files will be produced for the following analysis:
Condition GF upregulation
Condition GF downregulation
Condition SPF upregulation
Condition SPF downregulation
}
