% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeARIs,BenchmarkMetrics-method}
\alias{ComputeARIs,BenchmarkMetrics-method}
\title{Compute ARIs for BenchmarkMetrics objects}
\usage{
\S4method{ComputeARIs}{BenchmarkMetrics}(
  obj,
  labels,
  hclust_method = "complete",
  distance_type = "euclidean",
  sample_fraction = 1,
  num_cross_validation = 1,
  ...
)
}
\arguments{
\item{obj}{a BenchmarkMetrics object.}

\item{labels}{a character vector of variables you want to compute ARI for.}

\item{hclust_method}{a string indicating the method used for hierarchical clustering.}

\item{distance_type}{a string indicating what distance to use for hierarchical clustering}

\item{sample_fraction}{a number from 0 to 1 indicating how much the data to subset. (Improves speed for large datasets).}

\item{num_cross_validation}{an integer specifying how fold cross validation to do. Default is 1 (no cross validation).}

\item{...}{Additional params for other methods}
}
\value{
a BenchmarkMetrics object with computed ARIs under the ARI slot.
}
\description{
Computes Adjusted Rand Index for a specific categorical variable, using the 
fastcluster's fast hierarchical clustering implementation.
}
