% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeAssessments}
\alias{ComputeAssessments}
\alias{ComputeAssessments,BenchmarkMetrics-method}
\title{Compute Assessments}
\usage{
ComputeAssessments(obj, ...)

\S4method{ComputeAssessments}{BenchmarkMetrics}(obj, variables)
}
\arguments{
\item{obj}{A BenchmarkMetrics object.}

\item{...}{Additional parameters for other methods (currently not used)}

\item{variables}{A vector of variable names to compute the assessments for.}
}
\value{
A BenchmarkMetrics object with the computed assessments added.
}
\description{
This function computes various assessments (LISI, Silhouette, ARI) for multiple 
variables in a BenchmarkMetrics object.
}
\section{Functions}{
\itemize{
\item \code{ComputeAssessments(BenchmarkMetrics)}: Method for BenchmarkMetrics objects

}}
\examples{
\dontrun{
# Assuming `bm` is a BenchmarkMetrics object and `vars` is a vector of variable names
bm <- ComputeAssessments(bm, vars)
}

}
