% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeMultipleLISI,BenchmarkMetrics,ANY,character,ANY-method}
\alias{ComputeMultipleLISI,BenchmarkMetrics,ANY,character,ANY-method}
\title{Compute LISI for mutliple datasets}
\usage{
\S4method{ComputeMultipleLISI}{BenchmarkMetrics,ANY,character,ANY}(obj, reductions, variables, metadata, metrics_obj)
}
\description{
Compute LISI for mutliple datasets
}
