% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeMultipleLISI}
\alias{ComputeMultipleLISI}
\title{Compute LISI for multiple datasets}
\usage{
ComputeMultipleLISI(obj, reductions, variables, metadata, metrics_obj, ...)
}
\description{
Compute LISI for multiple datasets
}
