% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeMultipleSilhouette,BenchmarkMetrics-method}
\alias{ComputeMultipleSilhouette,BenchmarkMetrics-method}
\title{Plot silhouette for multiple datasets}
\usage{
\S4method{ComputeMultipleSilhouette}{BenchmarkMetrics}(obj, variables, result_format, ...)
}
\description{
Plot silhouette for multiple datasets
}
