% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeMultipleSilhouette}
\alias{ComputeMultipleSilhouette}
\title{Plot silhouette for multiple datasets}
\usage{
ComputeMultipleSilhouette(obj, ...)
}
\description{
Plot silhouette for multiple datasets
}
