% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations.R
\name{confALK}
\alias{confALK}
\title{confALK}
\usage{
confALK(species, quarter)
}
\arguments{
\item{species}{The species of interest.}

\item{quarter}{The quarter of interest.}
}
\value{
A dataframe with configurations needed for constructing the ALK, i.e. maximum age, minimum length and maximium length.
}
\description{
returns Configurations needed for the ALK
}
