% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_var.R
\name{content_var}
\alias{content_var}
\title{Generate vector from text in "Content" field of ingested emails}
\usage{
content_var(pdfdata)
}
\arguments{
\item{pdfdata}{email data as a list already ingested with pdfselec()}
}
\value{
character vector representing "Content" field of email data
}
\description{
Emails saved as one pdf file are ingested using pdfselect(). This helper function selects out the text
of each email in the "Content" field (the main text of the email) and generates a vector from this so that the content of each
email is represented.
}
