% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Correlation analysis plot}
\usage{
correlation_analysis(
  data,
  break_length = 5,
  length_lim = c(0, 1),
  number_digits = 3
)
}
\arguments{
\item{data}{A matrix for analyzing correlation.}

\item{break_length}{Number of segments divided into legend}

\item{length_lim}{Legend range}

\item{number_digits}{Reserved decimal places}
}
\value{
result
}
\description{
Correlation analysis plot
}
\details{
Title
}
\examples{
correlation_analysis(mtcars)
}
\author{
Fei Liang
}
