% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{CPUElength}
\alias{CPUElength}
\title{CPUElength}
\usage{
CPUElength(RFA, species, year, quarter, dat, bootstrapProcedure = "simple",
  B = 10)
}
\arguments{
\item{RFA}{The roundfish area of interest.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{bootstrapProcedure}{The bootstrap procedure ("simple", ...)}

\item{B}{The number of simulations in the selected bootstrap procedure}
}
\value{
Returns the estimated mCPUE per length class in the given RFA with uncertainty
}
\description{
.
}
