% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{CPUErfa}
\alias{CPUErfa}
\title{CPUErfa}
\usage{
CPUErfa(RFA, species, year, quarter, dat, bootstrapProcedure = "datras",
  B = 10, ALKprocedure = "datras", doBootstrap = TRUE, fit = NULL,
  report = NULL, lengthDivision)
}
\arguments{
\item{RFA}{The roundfish area of interest.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{bootstrapProcedure}{The bootstrap procedure ("simple", "stratisfied", ...)}

\item{B}{The number of simulations in the selected bootstrap procedure}

\item{ALKprocedure}{Either datras, haulbased or modelbased.}

\item{doBootstrap}{A boolean stating if bootstrap should be done. Default is TRUE.}
}
\value{
Returns the mCPUE per age class in the given RFA with uncertainty
}
\description{
.
}
