% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_years}
\alias{default_years}
\title{Dummy vector of year numbers of length 7}
\format{
An object of class \code{integer} of length 7.
}
\usage{
default_years
}
\description{
Dummy vector of year numbers of length 7
}
\keyword{datasets}
