% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineProgressBar.R
\name{defineProgressBar}
\alias{defineProgressBar}
\title{Define a progress bar}
\usage{
defineProgressBar(n_iterations)
}
\arguments{
\item{n_iterations}{The number of iterations for the progress bar.}
}
\value{
The progress bar object.
}
\description{
This function creates a progress bar using the \code{txtProgressBar} function.
}
