% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Sample Data for Demostration}
\format{
An object of class \code{data.frame} with 654 rows and 6 columns.
}
\usage{
data(demo_data)
}
\description{
A common measurement of lung function is the forced expiratory
volume FEV, which measures how much air a person can blow out of
the lungs in a short period of time. A higher FEV is usually
associated with better respiratory function. Data on 654
children are available for the following covariates:
\itemize{
   \item{id} - {Observation ID}
   \item{A} - {AGE (years)}
   \item{H} - {HEIGHT (inches)}
   \item{M} - {MALE (M=0 for female, 1 for male)}
   \item{S} - {SMOKE(S=0 for non-smoker and 1 for smoker)}
   \item{fev} - {forced expiratory volume FEV1}
}
}
\references{
Module K, BIOSTAT650, Dr. Xu Shi ,University of Michigan, School of Public Health
}
\keyword{datasets}
