% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamond plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{string name for var x in datafram df}

\item{vary}{string name for var y in dataframe df}
}
\value{
a plot with diamond markers \code{vary} versus \code{varx}
}
\description{
this funciton producs a ggplot with \code {vary} versus \code {varx} with diamond markers
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))

diamond_plot(df.test, 'x', 'y')


}
