% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometric_linear_regression.R
\name{econometric_linear_regression}
\alias{econometric_linear_regression}
\title{Perform Linear Regression}
\usage{
econometric_linear_regression(formula, data)
}
\arguments{
\item{formula}{A formula specifying the regression model.}

\item{data}{A data frame containing the variables.}
}
\value{
A lm object.
}
\description{
Perform Linear Regression
}
\examples{
\dontrun{
my_model <- econometric_linear_regression(Sepal.Length ~ Sepal.Width + Petal.Length, iris)
}
}
