% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaillist.R
\docType{data}
\name{emaillist}
\alias{emaillist}
\title{List of email data}
\format{
list
}
\source{
generated from testdata.pdf
}
\usage{
emaillist
}
\description{
This is a list created from running the pdfselect() function on testdata.pdf.
}
\keyword{datasets}
