% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailPDFdf.R
\name{emailPDFdf}
\alias{emailPDFdf}
\title{Ingest pdf file of emails to create data frame with required variables}
\usage{
emailPDFdf(pdfinput)
}
\arguments{
\item{pdfinput}{a pdf file of received emails}
}
\value{
a data frame containing subject, date, from, to and content of emails
}
\description{
Emails that are saved as a single pdf file are ingested and ultimately converted to
a data frame of variables "subject", "date", "to", "from", "content" with each row
representing each email received
}
