% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{print environment info}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{}
}
\value{
The environment info and  a message
}
\description{
print environment info
}
\examples{
environment_info('hi')
}
