% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate}
\alias{estimate}
\title{Ultra fast equivalent of fonction coeff()}
\usage{
estimate(y, x)
}
\arguments{
\item{y}{observations}

\item{x}{regressors}
}
\value{
Estimated coefficients of linear regression, using least square loss
}
\description{
Mimicks coeff() function but with greater SWAG
}
\examples{
x <- seq(0, 1, length = 10);
y <- 1 + 2 * x + rnorm(length(x), 0, 0.1)
estimate(y, x)
}
\seealso{
lm function (really not as good)
}
