% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate OLS coefficients by linear algebra}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numeric vector of outcomes.}

\item{X}{A numeric matrix of regressors.}
}
\value{
A numeric vector of estimated coefficients.
}
\description{
Estimate OLS coefficients by linear algebra
}
\examples{
y <- c(1, 2, 3)
X <- cbind(1, c(1, 2, 3))
estimate_beta(y, X)
}
