% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Title euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
number
}
\description{
Title euclidean function
}
\examples{
euclidean(45, 10)
}
