% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{example}
\usage{
example(filepath)
}
\description{
example
}
\examples{
add_lgd <- function(){
rast <- as.raster(hcl.colors(100))
usr <- par("usr")
rasterImage(rast,
            xleft = usr[2],
            ybottom = usr[3],
            xright = usr[2] + .2,
            ytop = usr[4],
            xpd = TRUE)
}

par(mfrow = c(2,2))
plot(1,1, main = "default")
add_lgd()

plot(1,1, xlim = c(0,1), ylim = c(0,1), asp = 1, main = "xlim = c(0,1), ylim = c(0,1), asp = 1")
add_lgd()

plot(1,1, xlim = c(0,1), asp = 1, main = "xlim = c(0,1), asp = 1")
add_lgd()

plot(1,1, ylim = c(0,1), asp = 1, main = "ylim = c(0,1), asp = 1" )
add_lgd()

}
