% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_HSI_dat.R
\docType{data}
\name{example_HSI_dat}
\alias{example_HSI_dat}
\title{Fake HSI Data}
\format{
An object of class \code{data.frame} with 10202 rows and 3 columns.
}
\usage{
data(example_HSI_dat)
}
\description{
Longer more detailed description of this Fake HSI Data
}
\examples{
data(example_HSI_dat)
BottomType_HSI <- CalcHSI_cat(DataInput = example_HSI_data, EnvVariable = Bottom)
}
\keyword{datasets}
