% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_len_dat.R
\docType{data}
\name{example_len_dat}
\alias{example_len_dat}
\title{Fake FWC Length Frequency Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10508 rows and 5 columns.
}
\usage{
data(example_len_dat)
}
\description{
Longer more detailed description of this fake FWC Length Frequency Data
}
\examples{
data(example_len_dat)
Make_LF_Plot(len_dat= example_len_dat, species = "Lutjanus griseus")
}
\keyword{datasets}
