% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_function.R
\name{f_add}
\alias{f_add}
\title{Title add function}
\usage{
f_add(x, y)
}
\arguments{
\item{x}{first input number}

\item{y}{second input number}
}
\value{
the sum
}
\description{
Title add function
}
\examples{
{f_add(3,9)}

}
