% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy-function.R
\name{fancy_function}
\alias{fancy_function}
\title{Some fancy stuff}
\usage{
fancy_function(data, formula, ...)
}
\arguments{
\item{data}{\code{data.frame}}

\item{formula}{A formula involving variables in data}

\item{...}{further unused arguments.}
}
\value{
A model frame including all predictors on the right-hand side of \code{formula}.
}
\description{
Wowzers
}
\examples{
fancy_function(data = mtcars, formula = cyl ~ disp)
}
