% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state code. If it does not match the state code, the function shows "invalid STATE number:" message.}

\item{year}{The selected year to be studied.}
}
\value{
Returns a map shows the position of accidents.
}
\description{
This function plots the position of accidents in a map with corresponding year and state from user input.
}
\examples{
fars_map_state(01, 2013)

}
