% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of file to read.}
}
\value{
The data is returned as a data frame.
}
\description{
This function reads csv data file if it exists and returns as a data frame.
If not, it shows an error message.
}
