% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Read fars_data based on years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Years to read in a vector of numbers.}
}
\value{
This function returns a list of data. The each element of the list
is the data from "accident_(year).csv.bz2". The file name to read is
generated by make_filename function and read by fars_read function.
If the year specified by the user does not have a file, this function
gives a "invalid year:" warning and returns NULL.
}
\description{
Returns data of specific years from US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System.
}
