% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars_data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years to be read.}
}
\value{
This function returns a data frame which shows the number of accidents in each month and year.
}
\description{
This function returns a data frame which has months in row, years in
column, and the number of corresponding accidents number as the value of the data frame.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)

}
