% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Combine levels of two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{a factor}

\item{b}{a factor}
}
\value{
return a factor with the levels of a and b
}
\description{
Combine levels of two factors
}
