% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{FilterLowlyExpressedGenes}
\alias{FilterLowlyExpressedGenes}
\title{Filter out lowly expressed genes using Seurat's variance stabilizing transformation}
\usage{
FilterLowlyExpressedGenes(obj, batch_variable, ngenes_per_batch = 10000, ...)
}
\description{
Filter out lowly expressed genes using Seurat's variance stabilizing transformation
}
