% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_FL_Places.R
\docType{data}
\name{flPlacesRace}
\alias{flPlacesRace}
\title{The ACS "Places" Data Set for Florida}
\format{An object of class \code{acs} with 918 rows and 10 columns.}
\usage{
flPlacesRace
}
\description{
The ACS "Places" Data Set for Florida
}
\keyword{datasets}
