% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_img.R
\name{get_img}
\alias{get_img}
\title{Create a graphical object (GROB) from png or jpg}
\usage{
get_img(filename)
}
\arguments{
\item{filename}{Filepath to png or jpg file.}
}
\value{
A \code{GROB} that can be added to a ggplot.
}
\description{
Create a graphical object (GROB) from png or jpg
}
