% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_mot_recent_date}
\alias{get_mot_recent_date}
\title{Get date of most recent day of week
This function gets the most recent date for a day of the week}
\usage{
get_mot_recent_date(day)
}
\arguments{
\item{day}{character string in format of "Monday", "Tuesday",etc.}
}
\value{
date object
}
\description{
Get date of most recent day of week
This function gets the most recent date for a day of the week
}
\examples{
get_mot_recent_date("Monday")
}
