% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{getEstimatesCPUEage}
\alias{getEstimatesCPUEage}
\title{getEstimateCPUEage}
\usage{
getEstimatesCPUEage(RFA, species, year, quarter, dataHL, dataCA,
  percentOfAreaRepresentative = NULL, bootstrapProcedure = "simple",
  B = 10, removeProportionsOfCA = 0, removeProportionsOfHL = 0,
  procedure = "")
}
\arguments{
\item{RFA}{The roundfish area of interest.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{dataHL}{datras-data with length}

\item{dataCA}{datras-data with age}

\item{percentOfAreaRepresentative}{the percentage of the statical recangle within sea depth intervall}

\item{bootstrapProcedure}{The bootstrap procedure ("simple", "stratisfied", ...)}

\item{B}{The number of simulations in the selected bootstrap procedure}

\item{newPeocedure}{Logical if we use the new procedure (defaul = false)}
}
\value{
Returns the mCPUE per age class in the given RFA with uncertainty
}
\description{
.
}
