% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetNGrams.R
\name{GetNGrams}
\alias{GetNGrams}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
GetNGrams(words, n)
}
\arguments{
\item{words}{A vector of strings that represents a tokenized document}

\item{n}{The size of the n-gram window}
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}

