% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotaSamplingCheck.R
\name{getStationsWithAge}
\alias{getStationsWithAge}
\title{Get stations with age measurement for the given species and length group}
\usage{
getStationsWithAge(hh, ca, species, lngtCM)
}
\arguments{
\item{hh}{the HH table for the survey}

\item{ca}{the CA table for the survey}

\item{lngtCM}{the lengthgroup as identified by column lngtCM in CA and HL}

\item{scientific}{name of species}
}
\value{
data frame formatted as hh
}
\description{
Get stations with age measurement for the given species and length group
}
