% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotaSamplingCheck.R
\name{getStationsWithLength}
\alias{getStationsWithLength}
\title{Get stations with length measurement for the given species and length group}
\usage{
getStationsWithLength(hh, hl, species, lngtCM)
}
\arguments{
\item{hh}{the HH table for the survey}

\item{hl}{the HL table for the survey}

\item{lngtCM}{the lengthgroup as identified by column lngtCM in CA and HL}

\item{scientific}{name of species}
}
\value{
data frame formatted as hh
}
\description{
Get stations with length measurement for the given species and length group
}
