% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe-arithmetic.R
\name{\%plus\%}
\alias{\%plus\%}
\title{Add two numbers with an operator}
\usage{
x \%plus\% y
}
\arguments{
\item{x, y}{Numbers}
}
\value{
The sum of x and y
}
\description{
Add two numbers with an operator
}
\examples{
2 \%plus\% 2
}
