% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat_matrix.R
\name{hat_matrix}
\alias{hat_matrix}
\title{hat_matrix}
\usage{
hat_matrix(fit.model)
}
\arguments{
\item{fit.model}{takes an R object, returned by lm}
}
\value{
the index of the maximum leverage  value
}
\description{
Finds the index of the largest leverage value
}
