% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print hello world}
\usage{
hello(x)
}
\arguments{
\item{x}{A character string to be concatencated in print statement}
}
\value{
Hello world
}
\description{
Prints hello world
}
\examples{

library(testpackage)
hello('string')

}
