% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyClass.R
\name{hello_world.MyClass}
\alias{hello_world.MyClass}
\title{A sample method}
\usage{
\method{hello_world}{MyClass}(x, ...)
}
\arguments{
\item{x}{class object}

\item{...}{other parameters}
}
\description{
A sample method
}
