% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Title
Creates a histogram}
\usage{
histogram(parametersfile)
}
\arguments{
\item{parametersfile}{json structure containing information to create a histogram}
}
\value{
a ggplot object
}
\description{
Title
Creates a histogram
}
