\name{custom_index}
\alias{custom_index}
\title{Custom Index of Functions}
\docType{package}
\description{
  This is a custom index for the package, organizing functions into categories.
}

\section{Main Functions}{
  \itemize{
    \item \code{\link{ComputeARIs}}: Computes ARIs
  }
}

\section{Helper Functions}{
  \itemize{
    \item \code{\link{ComputeAssessments}}: Computes assessments
  }
}
