% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_emissions_function.R
\name{join_emissions}
\alias{join_emissions}
\title{join_emissions uses dummy data}
\usage{
join_emissions(df)
}
\arguments{
\item{df}{The traffic data dataframe}
}
\value{
The joined dataset
}
\description{
Join dummy emissions to df
}
\examples{
df1 <- join_emissions(df = trafficdata1);
df2 <- join_emissions(df = trafficdata2);
}
