% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot.R
\name{kmeans_plot}
\alias{kmeans_plot}
\title{Title}
\usage{
kmeans_plot(
  data,
  centers,
  row,
  col,
  center_col = "blue",
  line_col = "grey",
  las = 2
)
}
\arguments{
\item{data}{matrix}

\item{centers}{Number of clusters}

\item{row}{Number of drawing columns line}

\item{col}{Number of drawing columns}

\item{center_col}{Color of center line}

\item{line_col}{Color of line}

\item{las}{Whether to rotate the abscissa, 1 means no rotation and 2 means rotation.}
}
\value{
data_new2
}
\description{
Title
}
\examples{
runif(10)
}
