% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot3.R
\name{kmeans_plot3}
\alias{kmeans_plot3}
\title{Title}
\usage{
kmeans_plot3(
  data,
  centers,
  row,
  col,
  center_col = "black",
  las = 2,
  box = TRUE,
  point_plot = FALSE
)
}
\arguments{
\item{data}{matrix}

\item{centers}{Number of clusters}

\item{row}{Number of drawing columns line}

\item{col}{Number of drawing columns}

\item{center_col}{Color of center line}

\item{las}{Whether to rotate the abscissa, 1 means no rotation and 2 means rotation.}

\item{box}{Logic value, whether to close the graph.}

\item{point_plot}{Logic value,whether to add points in the diagram.}
}
\value{
data_new2
}
\description{
Title
}
\examples{
kmeans_test_data <- kmeans_test_data
kmeans_test_group <- kmeans_test_group
data <- testpackage::means_calculate(kmeans_test_data,
kmeans_test_group)
testpackage::kmeans_plot(data,
                         centers = 6,
                         row = 2,
                         col = 3,
                         center_col = "blue",
                         las = 2)
                         testpackage::kmeans_plot2(data,
                         centers = 6,
                         row = 2,
                         col = 3,
                         center_col = "blue",
                         las = 2
                         )
testpackage::kmeans_plot3(data,
                         centers = 6,
                         row = 2,
                         col = 3,
                         center_col = "blue",
                         las = 2
                         )
testpackage::kmeans_plot3(data = data,
                         centers = 6,
                         row = 2,
                         col = 3,
                         center_col = "blue",
                         las = 2,
                         box = FALSE,
                         point_plot = TRUE
                         )

}
