% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot4.R
\name{kmeans_plot4}
\alias{kmeans_plot4}
\title{Title}
\usage{
kmeans_plot4(data, centers, row, col, center_col = "black", las = 2)
}
\arguments{
\item{data}{matrix}

\item{centers}{Number of clusters}

\item{row}{Number of drawing columns line}

\item{col}{Number of drawing columns}

\item{center_col}{Color of center line}

\item{las}{Whether to rotate the abscissa, 1 means no rotation and 2 means rotation.}
}
\value{
data_new2
}
\description{
Title
}
\examples{
runif(10)
}
