% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot_old.R
\name{kmeans_plot_old}
\alias{kmeans_plot_old}
\title{Generate a time series of fractional Brownian motion.}
\arguments{
\item{data}{Expression matrix}

\item{group}{Sample group}

\item{cluster_num}{Number of clusters}

\item{row}{plot number of row}

\item{col}{plot number of col}
}
\description{
This function generatea a time series of one dimension fractional Brownian motion.
adapted from http://www.mathworks.com.au/matlabcentral/fileexchange/38935-fractional-brownian-motion-generator .
}
\details{
Big data trend analysis
}
\examples{
runif(10)
}
