% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_of_kmeans_test_data.R
\docType{data}
\name{kmeans_test_data}
\alias{kmeans_test_data}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{S1-1}{sample1}
  \item{S1-2}{sample2}
  \item{S1-3}{sample1}
  \item{S1-4}{sample2}
  \item{S1-5}{sample1}
  \item{S1-6}{sample2}
  \item{S2-1}{sample1}
  \item{S2-2}{sample2}
  \item{S2-3}{sample1}
  \item{S2-4}{sample2}
  \item{S2-5}{sample1}
  \item{S2-6}{sample2}
  \item{S3-1}{sample1}
  \item{S3-2}{sample2}
  \item{S3-3}{sample1}
  \item{S3-4}{sample2}
  \item{S3-5}{sample1}
  \item{S3-6}{sample2}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
kmeans_test_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
