% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/littleforecast.R
\name{littleforecast}
\alias{littleforecast}
\title{Forecast with Arima Model}
\usage{
littleforecast(data, nyears = 10)
}
\arguments{
\item{data}{A vector (time series) of data}

\item{nyears}{Number of time steps to forecast forward}
}
\value{
A plot of a forecast.
}
\description{
This fits an Arima model to data with forecast's auto.arima() function and plots
a forecast with the forecast() function.
}
\examples{
dat <- WWWusage
littleforecast(dat, nyears=100)
}
