% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{load titanic.csv}
\usage{
load(file_path, output_path)
}
\arguments{
\item{file_path}{the data}

\item{output_path}{the output directory}
}
\value{
the cleaned data
}
\description{
load titanic.csv
}
\examples{
\dontrun{
load("data/raw/titanic.csv", "output/cleaned.csv")
}
}
