% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Make a file name with year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A specific year}
}
\value{
This function returns a string of file name.
}
\description{
Returns a file name "accident_year.csv.bz2" with year is specified by a
user (using the \code{year} argument).
}
