% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_LF_Plot.R
\name{Make_LF_Plot}
\alias{Make_LF_Plot}
\title{Length Frequency Graph w/Summary Stats}
\usage{
Make_LF_Plot(len_dat, species)
}
\arguments{
\item{len_dat}{dataframe of FWC length frequency data. Has columns "Scientificname", "Commonname", "sl" (Standard Length), "COUNT", and "Gear"}

\item{species}{a string representing the species you want to plot}
}
\value{
a length frequency histogram plot and a summary table
}
\description{
This makes a length-frequency graph for any particular species, with a summary data table placed below the graph. It works off of a previously made dataset with columns for Scientificname", "Commonname", "sl" (Standard Length), "COUNT", and "Gear"
}
\examples{
data(example_len_dat)
Make_LF_Plot(len_dat= example_len_dat, species = "Lutjanus griseus")
}
