% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sppcode.r
\name{make_sppcode}
\alias{make_sppcode}
\title{make_sppcode}
\usage{
make_sppcode(data, sppname)
}
\arguments{
\item{data}{Name of data frame that contains at least one column with Latin names}

\item{sppname}{Quoted name of the column that contains the Latin names}
}
\value{
Returns a data frame with a new column named sppcode.
}
\description{
Make a 6-letter code with first 3 letters of genus and species
}
\examples{
example_dat <- data.frame(Latin_Name = c("Carex limosa", "Arethusa bulbosa",
     "Malaxis unifolia", "Calopogon tuberosus"),
cover = c(10, 40, 10, 50),
stems = c(50, 20, 10, 10))
head(example_dat)
example_dat2 <- make_sppcode(example_dat, "Latin_Name")
head(example_dat2)
}
