% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebarplot.R
\name{makebarplot}
\alias{makebarplot}
\title{Make a simple bar plot}
\usage{
makebarplot(mydf, ...)
}
\arguments{
\item{mydf}{A dataframe with an X and a Y column}
}
\value{
A bar plot in ggplot ready to plot
}
\description{
Make a simple bar plot
}
\examples{
makebarplot(mydataframe)
}
