% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_waffle.R
\name{MakeWaffle}
\alias{MakeWaffle}
\title{Make a Waffle Plot}
\usage{
MakeWaffle(vec, title, colours = NULL)
}
\arguments{
\item{vec}{a numeric vector of strictly non-negative values}

\item{title}{the title of the plot}

\item{colours}{a character vector of colours the same length as \code{vec}}
}
\value{
Nothing. A waffle plot is displayed
}
\description{
Make a Waffle Plot
}
\examples{
  MakeWaffle(vec = runif(5), title = "test")
}
