% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchDifficultyCoefficient.R
\name{match_diff_coeff}
\alias{match_diff_coeff}
\title{Match Difficulty Coefficient function}
\usage{
match_diff_coeff(current_rank, match_matrix)
}
\description{
This function assigns a score evaluating the current form 
of all teams based on the first ten results of the current season.
  
[INSERT MOTIVATION OF DIFFICULTY COEFF COMPUTATION]
 
Formula: 
[INSERT FORMULA]
}
\examples{
match_diff_coeff()
}

