% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxTemp.R
\docType{data}
\name{MaxTemp}
\alias{MaxTemp}
\title{A time series of daily maximum temperatures in Flagstaff, AZ. The data were
downloaded from www.ncdc.noaa.gov.}
\format{a data frame with 10882 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{MaxTemp}{Daily maximum temperature in degrees Farhenheit.}
}}
\usage{
MaxTemp
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ. The data were
downloaded from www.ncdc.noaa.gov.
}
\keyword{datasets}
