% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_calculate.R
\name{means_calculate}
\alias{means_calculate}
\title{means calculate}
\usage{
means_calculate(data, group)
}
\arguments{
\item{data}{Multiomics data}

\item{group}{Sample grouping data}
}
\value{
all_sample_mean
}
\description{
means calculate
}
\details{
Title
}
\author{
Fei Liang
}
