% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{miss}
\alias{miss}
\title{Deal with missing values}
\usage{
miss(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
a list
}
\description{
Check if there are any missing values.
Return total and percentage of missing values per column.
Delete missing values and return complete data set.
}
