% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyClass.R
\name{my_class}
\alias{my_class}
\title{Creates a new model}
\usage{
my_class(x)
}
\arguments{
\item{x}{a character}
}
\value{
a new model
}
\description{
Creates a new model
}
\examples{
a <- my_class("abcd")
}
