% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fxn.R
\name{my_fxn}
\alias{my_fxn}
\title{Print hello in various languages}
\usage{
my_fxn(language = "english")
}
\arguments{
\item{language}{Character string of language to use. Must be lowercase. Default is 'english'}
}
\value{
Character string
}
\description{
Print hello in various languages
}
\examples{
my_fxn(language = "english")
}
