% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_pkg_data}
\alias{my_pkg_data}
\title{Fake data to test package}
\format{
\subsection{'chk'}{

A data frame with 1000 rows and 1 variable:
\describe{
\item{my_pkg_data}{This is my package data}
}
}
}
\usage{
my_pkg_data
}
\description{
This is fake data to test an R package.
}
\keyword{datasets}
