% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot theme
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
    ggplot2::geom_point() +
    my_theme()
}
}
