% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyClass.R
\name{MyClass}
\alias{MyClass}
\title{A Test R6 Class}
\description{
A Test R6 Class

A Test R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{a}}{something}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MyClass-new}{\code{MyClass$new()}}
\item \href{#method-MyClass-hello_world}{\code{MyClass$hello_world()}}
\item \href{#method-MyClass-summary}{\code{MyClass$summary()}}
\item \href{#method-MyClass-simulate}{\code{MyClass$simulate()}}
\item \href{#method-MyClass-clone}{\code{MyClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-new}{}}}
\subsection{Method \code{new()}}{
initializes the myclass object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{object for a}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-hello_world"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-hello_world}{}}}
\subsection{Method \code{hello_world()}}{
prints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$hello_world()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-summary"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-summary}{}}}
\subsection{Method \code{summary()}}{
summarizes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-simulate}{}}}
\subsection{Method \code{simulate()}}{
summarizes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$simulate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MyClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MyClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
