% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r, R/myfunc.r
\name{myfunc}
\alias{myfunc}
\title{My first function}
\usage{
myfunc(a, b)

myfunc(a, b)
}
\arguments{
\item{a}{A numerical vector.}

\item{b}{Also a numerical vector.}

\item{y}{an n x 1 vector of observations}

\item{X}{an n x k vector of observations}
}
\value{
beta_hat An estimate of beta in y = beta X + epsilon

A numerical vector of a + b * a.
}
\description{
My first function

My first function
}
\examples{
myfunc(3, 5)
myfunc(3, 5)
}
