% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctest.r
\name{myfunctest}
\alias{myfunctest}
\title{My second function}
\usage{
myfunctest(a, b)
}
\arguments{
\item{a}{A numerical vector.}

\item{b}{Also a numerical vector.}
}
\value{
A numerical vector of a + b * a.
}
\description{
My second function
}
\examples{
myfunctest(3, 5)
}
