% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_functions.R
\name{neural_net}
\alias{neural_net}
\title{Simple Neural Net}
\usage{
neural_net(y, X, init.val = rep(0, dim(X)[2] + 1))
}
\arguments{
\item{y}{A binary \code{vector} representing the output.}

\item{X}{A \code{matrix} of inputs.}

\item{init.val}{A \code{numeric} vector used for starting values (defaults to vector of zeros).}
}
\value{
A \code{vector} containing the estimated weights.
}
\description{
Compute weights and bias of a simple neural net.
}
\examples{
set.seed(1)
gender <- sample(c(0,1), size = 100, replace = TRUE) # x1 (input)
age <- round(runif(100, 18, 80)) # x2 (input)
xb <- 3.5*gender + 0.2*age - 9 # w1*x1 + w2*xw + b
p <- 1/(1 + exp(-xb))
y <- rbinom(n = 100, size = 1, prob = p) # output
regr <- cbind(gender, age) # create matrix of inputs
weights <- neural_net(y = y, X = regr)
}
\author{
Rob
}
