% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{NormalizeRNA}
\alias{NormalizeRNA}
\alias{NormalizeRNA,BenchmarkMetrics-method}
\title{Run all current popular RNA normalizations}
\usage{
NormalizeRNA(obj, batch_variable = "batch", num_pcs = 30, ...)

\S4method{NormalizeRNA}{BenchmarkMetrics}(obj, batch_variable = "batch", num_pcs = 30, ...)
}
\arguments{
\item{obj}{A BenchmarkMetrics object with populated raw_data and metadata}

\item{batch_variable}{A string indicating which metadata column corresponds to batch information}

\item{num_pcs}{An integer indicating how many PCs to compute for dim reduction}

\item{...}{Additional arguments for other methods (currently not used)}
}
\value{
A BenchmarkMetrics object with the adjusted data, their PCs and runtimes added to the corresponding slots.
}
\description{
Runs Seurat logNormalized, Seurat CCA, Seurat SCT, fastMNN and Harmony, 
also computes their PCAs and records runtimes. All stored back into the BenchmarkMetrics object.
}
\section{Functions}{
\itemize{
\item \code{NormalizeRNA(BenchmarkMetrics)}: S4 method for class BenchmarkMetrics

}}
