% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_influence.R
\name{outlier_influence}
\alias{outlier_influence}
\title{outlier_influence}
\usage{
outlier_influence(input_data, fit.model,
  option = c("dffits", "cd" , "cvr"),
    high.influence = FALSE)
}
\arguments{
\item{input_data}{a data frame containing the variables in th model}

\item{fit.model}{takes an R object, returned by \link[stats]{lm}}

\item{option}{which measure of influence diagnostics should be computed.
"dffits" is for DFFITS, "cd" is for Cook's distance,
"cvr" is for COVRATIO}

\item{high.influence}{(Default)FALSE; if TRUE, influential observation along with their
respective diagnostic statistics will be returned.}
}
\value{
If high.influence is set TRUE, each diagnostic statistic will return
a list containing two elements.
\itemize{
  \item{dffits, Cook's Distance, COVRATIO} - {depends on which option is chosen}
  \item{outliers} - {which observations are considered influential}
}
}
\description{
outlier_influence computes three diagnostic statistics for measuring influential points:
\describe{
  \item{DIFFTS}{Difference in the fitted value when ith observation is removed}
  \item{Cook's Distance}{shows the influence of each observation on the fitted response values}
  \item{COVRATIO}{influence on variance-covariance matrix}
}
}
\examples{
##Use demo_data (See ?demo_data for more information about the dataset)
m = lm(fev~A+H+M+S, data = demo_data)
outlier_influence(demo_data, m, option = c("dffits"))
outlier_influence(demo_data, m, option = c("cd"))
outlier_influence(demo_data, m, option = c("cvr"), high.influence = TRUE)

}
