% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA2.R
\name{PCA2}
\alias{PCA2}
\title{Title}
\usage{
PCA2(
  data,
  group,
  col = c("#00f000", "#0000f0", "#b30000"),
  size = 3,
  points_type = 15:17,
  line_type = 1,
  title = "Principal Component Analysis",
  xlim_min = 1.5,
  xlim_max = 2,
  ylim_min = 1.1,
  ylim_max = 1.5,
  legend = TRUE,
  legend_group = c("A", "B", "C "),
  edge_number_of_ellipse = 10
)
}
\arguments{
\item{data}{Sample matrix}

\item{group}{Sample group}

\item{col}{dot and ellipse color.}

\item{size}{point size}

\item{points_type}{points_type}

\item{line_type}{ellipse line_type}

\item{title}{title}

\item{xlim_min}{Axis margin}

\item{xlim_max}{Axis margin}

\item{ylim_min}{Axis margin}

\item{ylim_max}{Axis margin}

\item{legend}{Logic value, whether to display legend or not.}

\item{legend_group}{The name of the group in the legend}

\item{edge_number_of_ellipse}{Number of confidence interval edges}
}
\value{
pca_data
}
\description{
pca plot
}
\examples{
pca_test_data <- pca_test_data
pca_test_group <- pca_test_group
testpackage::PCA2(data = pca_test_data,
 group = pca_test_group)
pca_test_data <- pca_test_data[, c(1:8, 17:24)]
pca_test_group <- pca_test_group[c(1:8, 17:24), ]
testpackage::PCA2(data = pca_test_data,
group = pca_test_group,
col = c("#0000f0", "#b30000"),
legend_group = c("B", "C"),
edge_number_of_ellipse = 4,
size = 2,
xlim_min = 2,
ylim_min = 1.3,
ylim_max = 0.7)
testpackage::PCA2(data = pca_test_data,
group = pca_test_group,
col = c("#0000f0", "#b30000"),
legend_group = c("B", "C"),
edge_number_of_ellipse = 3,
size = 2,
xlim_min = 2,
ylim_min = 1.3,
ylim_max = 0.7)
}
\author{
Fei Liang
}
