% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_of_pca_test_data.R
\docType{data}
\name{pca_test_data}
\alias{pca_test_data}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{A1}{sample1}
  \item{A2}{sample1}
  \item{A3}{sample1}
  \item{A4}{sample1}
  \item{A5}{sample1}
  \item{A6}{sample1}
  \item{A7}{sample1}
  \item{A8}{sample1}
  \item{B1}{sample1}
  \item{B2}{sample1}
  \item{B3}{sample1}
  \item{B4}{sample1}
  \item{B5}{sample1}
  \item{B6}{sample1}
  \item{B7}{sample1}
  \item{B8}{sample1}
  \item{C1}{sample1}
  \item{C2}{sample1}
  \item{C3}{sample1}
  \item{C4}{sample1}
  \item{C5}{sample1}
  \item{C6}{sample1}
  \item{C7}{sample1}
  \item{C8}{sample1}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
pca_test_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
