% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_of_pca_test_group.R
\docType{data}
\name{pca_test_group}
\alias{pca_test_group}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Sample}{Sample}
  \item{Group}{Group}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
pca_test_group
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
