% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfselect.R
\name{pdfselect}
\alias{pdfselect}
\title{Ingest pdf file of emails and create a list with required content}
\usage{
pdfselect(pdfinput)
}
\arguments{
\item{pdfinput}{a pdf file of received emails}
}
\value{
a list containing the selected email data
}
\description{
Emails saved as a single pdf file are read into R as a list using this function. Only text that is
relevant for creating a data frame of variables "subject", "date", "to", "from", "content"
is selected to return the final desired list.
}
