% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plifetime.R
\name{plifetime}
\alias{plifetime}
\title{Probability of Battery Lifetime}
\usage{
plifetime(a, b, c, d)
}
\arguments{
\item{a}{a continuous vertical line which intersects with lines 'c' and 'd'}

\item{b}{a continuous vertical line which intersects with lines 'c' and 'd'}

\item{c}{a continuous horizontal line which intersects with lines 'a' and 'b'}

\item{d}{a continuous horizontal line which intersects with lines 'a' and 'b'}
}
\value{
The probability that (\emph{X, Y}) falls within the rectangle, which is
created at the intersections of the the lines 'a', 'b', 'c', and 'd'.
}
\description{
Let X & Y be the survival times of two batteries. This functions
calculates the probability that (\emph{X, Y}) falls within the rectangle which
represents the distribution of battery lifetime.The modes of the survival time
are (x, y) = (sqrt(2)/2, sqrt(2)/2). X and Y denote the lifetimes of the batteries
of the component in standard units.
}
\details{
Below is the function's pdf:
```
pdf(x,y) = \{ 4xye^-(x^2 + y^2), when x > 0, y > 0
           \{ 0                  otherwise
```
This function is used to solve Exercise 2.1.6 on page 98 of the book.
}
\examples{
a <- 2
b <- 3
c <- 2
d <- 3

plifetime(a, b, c, d)

# Probability of (X, Y) surviving beyond the mode lifetime of the batteries
a <- 1
b <- 2
c <- Inf
d <- Inf

plifetime(a, b, c, d)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
