% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influential_outliers.R
\name{plot_dffits}
\alias{plot_dffits}
\title{Plot Influential Outliers}
\usage{
\itemize{
  \item{plot_dffit(fit.model)} - {plots DFFITS and labels the subjects that are outliers}
  \item{plot_cd(fit.model)}- {plots Cook's Distance and labels the subjects that are outliers}
  \item{plot_CVR(fit.model)} - {plots COVRATIO }
}
}
\arguments{
\item{fit.model}{an R object, returned by \link[stats]{lm}}
}
\description{
This suite of plotting functions are used to plot
DFFITS, Cook's Distance, and COVRATIO separately
}
\examples{
require(ggplot2)
m = lm(fev~A+H+M+S, data = demo_data)
plot_dffits(m)
plot_cd(m)
plot_CVR(m)

}
