% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iris.R
\name{plot_iris}
\alias{plot_iris}
\title{Plot iris data}
\usage{
plot_iris(dt, color)
}
\arguments{
\item{color}{color of the bars}

\item{x}{column that you want to plot}
}
\value{
data visualization
}
\description{
Plot iris data
}
