% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series}
\usage{
plot_timeseries(x)
}
\arguments{
\item{x}{numeric vector of length 7}
}
\value{
ggplot2 object
}
\description{
Plot time series
}
