% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotARIs,BenchmarkMetrics-method}
\alias{PlotARIs,BenchmarkMetrics-method}
\title{plot ARIs for BenchmarkMetrics objects}
\usage{
\S4method{PlotARIs}{BenchmarkMetrics}(obj, variable, title = "ARIs of different methods", ...)
}
\description{
plot ARIs for BenchmarkMetrics objects
}
