% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotbymonth}
\alias{plotbymonth}
\title{Plot ufo sightings by month}
\usage{
plotbymonth(state = "NV")
}
\arguments{
\item{state}{what state it's in,
Default is NV}
}
\value{
Returns a plot.  See \link{ufo}
}
\description{
Plot ufo sightings by month
}
