% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotHVG_Conservation,BenchmarkMetrics-method}
\alias{PlotHVG_Conservation,BenchmarkMetrics-method}
\title{Plot HVG conservation for BenchmarkMetrics onjects}
\usage{
\S4method{PlotHVG_Conservation}{BenchmarkMetrics}(
  obj,
  batch_variable = "batch",
  flavour = "vst",
  title = "HVG conservation plot"
)
}
\description{
Plot HVG conservation for BenchmarkMetrics onjects
}
