% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotMultipleLISI,BenchmarkMetrics,character-method}
\alias{PlotMultipleLISI,BenchmarkMetrics,character-method}
\title{Plot LISI for multiple datasets}
\usage{
\S4method{PlotMultipleLISI}{BenchmarkMetrics,character}(
  obj,
  variable,
  reductions,
  aspect_ratio = 1.3,
  title = NULL,
  levels = NULL
)
}
\description{
Plot LISI for multiple datasets
}
