% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotMultipleSilhouette,BenchmarkMetrics-method}
\alias{PlotMultipleSilhouette,BenchmarkMetrics-method}
\title{Plot silhouettes for multiple datasets}
\usage{
\S4method{PlotMultipleSilhouette}{BenchmarkMetrics}(
  obj,
  variable,
  plot_type = "violin",
  title = NULL,
  aspect_ratio = 1.3,
  ...
)
}
\description{
Plot silhouettes for multiple datasets
}
