% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotMultipleSilhouette}
\alias{PlotMultipleSilhouette}
\title{Plot silhouettes for multiple datasets}
\usage{
PlotMultipleSilhouette(obj, variable, plot_type, title, aspect_ratio, ...)
}
\description{
Plot silhouettes for multiple datasets
}
