% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotRuntime}
\alias{PlotRuntime}
\title{Plot runtimes stored in the metrics object}
\usage{
PlotRuntime(obj, title = "Runtime", log = F)
}
\description{
Plot runtimes stored in the metrics object
}
