% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSquare.R
\name{plotSquare}
\alias{plotSquare}
\title{Plot a square of side length r.}
\usage{
plotSquare(r = 1)
}
\arguments{
\item{r}{The side length of the square to
  be plotted}
}
\value{
None The function is called
  solely for its side effect
}
\description{
Normally, we would now describe what
the function is supposed to do
}
\examples{
plotSquare(5)
}
\seealso{
The general
  \code{\link[graphics]{plot}} function
}

