% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotaSamplingCheck.R
\name{plotStations}
\alias{plotStations}
\title{Plot stations color coded for missing age samples in specific length groups}
\usage{
plotStations(hh = HH, ca = CA, hl = HL, species = "Gadus morhua",
  lengthGroup = 38, polygons = rfa)
}
\arguments{
\item{hh}{HH table for the survey}

\item{ca}{CA table for the survey}

\item{hl}{HL table for the survey}

\item{polygons}{SpatialPolygonsDataFrame object for drawing on the map. Map be NULL.}

\item{scientific}{name of species}

\item{lngtCM}{the lengthgroup as identified by column lngtCM in CA and HL}
}
\description{
Plot stations color coded for missing age samples in specific length groups
}
