% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus1.R
\name{plus1}
\alias{plus1}
\title{plus1}
\usage{
plus1(x)
}
\arguments{
\item{x}{A numeric value or vector to which 1 will be added.}
}
\value{
A numeric value or vector with 1 added to each element of the input.
}
\description{
\code{plus1} adds 1 to the input value.
}
\examples{
plus1(3) # returns 4
plus1(c(1, 2, 3)) # returns c(2, 3, 4)
}
