% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLigue1.R
\name{predict_L1}
\alias{predict_L1}
\title{Predict Ligue 1 Football championship function}
\usage{
predict_L1()
}
\description{
This is the master function of our package that will call all other functions and data
needed to output a prediction standings of the French football championship using previous season standings
and the first ten matches of the season.
}
\examples{
predict_L1()
}

