% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{pretty_plot}
\alias{pretty_plot}
\title{Add a logo and text underneath a plot}
\usage{
pretty_plot(plot, source = "", logo_path, width = 4, text_size = 16)
}
\arguments{
\item{plot}{A \code{ggplot} to be amended.}

\item{source}{A character string of text to be added bottom left.}

\item{logo_path}{Filepath to png or jpg file of logo to be added bottom right.}

\item{width}{Size of the frame where text and logo are placed. Default is 4.}

\item{text_size}{Font size for the text. Default is 16}
}
\value{
A plot with text (left) and logo (right)
}
\description{
\code{pretty_plot} is a wrapper to place a logo at the bottom right and an optional text at the bottom left.
It is intended to give a consistent, 'corporate' feel to figures.
It adds an extra margin to the plot at the bottom where the logos and text are placed.
}
