% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priemel.R
\name{priemel}
\alias{priemel}
\title{Table of serum level and OV.BV ratio from Priemel et al 2010}
\format{This table is 675 by 2 \cr
each row correspond to serum level and ratio of osteoid volume (OV) to bone volume (BV). \cr
\strong{SerumLevelInNgPerMl} - serum level. \cr
\strong{OV.BV} - ratio of osteoid volume (OV) to bone volume (BV). \cr}
\source{
The physiologic data come from a study by \href{https://www.ncbi.nlm.nih.gov/pubmed/19594303}{Priemel et al (2010)}.
}
\description{
A dataset relating physiologic response to serum level
}
