% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printemoji.R
\name{print_emoji}
\alias{print_emoji}
\alias{print_emoji.data.frame}
\title{This function prints a Halloween-themed message.}
\usage{
print_emoji(x, ...)

\method{print_emoji}{data.frame}(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{arguments passed to \code{\link{print}}}
}
\description{
This function prints a Halloween-themed message.
}
\examples{
print_emoji(mtcars)
}
