% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{Title
Reads columns from a data table}
\usage{
read_data(filename, select_columns)
}
\arguments{
\item{filename}{a string filename including the relative}

\item{select_columns}{a vector including the column names from the data file}
}
\value{
if succeds, this function returns a data table
}
\description{
Additional info
}
