% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers from data}
\usage{
remove_outliers(data, min_val, max_val)
}
\arguments{
\item{data}{vector of data}

\item{min_val}{mimimun value allowed in dataset}

\item{max_val}{maximum value allowed in dataset}
}
\description{
Removes outliers from a vector of data based on a minimum and maximum threshold
}
