% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeDataDetailedCA}
\alias{removeDataDetailedCA}
\title{removeDataDetailedCA}
\usage{
removeDataDetailedCA(datDetailed, species, quarter, lengthDivision,
  samplesWithinEachIntervall)
}
\arguments{
\item{datDetailed}{The quarter of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
