% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeDataDetailedHL}
\alias{removeDataDetailedHL}
\title{removeDataDetailedHL}
\usage{
removeDataDetailedHL(datDetailed, removeProcedure, propRemove, RFA)
}
\arguments{
\item{datDetailed}{The quarter of interest.}

\item{removeProcedure}{Removal procedure, given as an integer. removeProcedure==1 means that data is removed at random.}

\item{propRemove}{proportion of data to remove.}

\item{RFA}{The roundfish area of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
