% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeObsFromHaul}
\alias{removeObsFromHaul}
\title{removeObsFromHaul}
\usage{
removeObsFromHaul(obsTmp, lengthDivision)
}
\arguments{
\item{datDetailed}{The quarter of interest.}

\item{removeProcedure}{Removal procedure, given as an integer. removeProcedure==1 means that data is removed at random.}

\item{RFA}{The roundfish area of interest.}

\item{propRemove}{proportion of data to remove.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
