% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_3.R
\name{res_3}
\alias{res_3}
\title{res_3}
\usage{
res_3(input_data, fit.model ,r = c("z", "int", "ex"))
}
\arguments{
\item{input_data}{a data frame containing the variables in th model}

\item{fit.model}{a R object, returned by \link[stats]{lm}}

\item{r}{what type of resisuals will be computed: "z" is standardized resisuals,
"int" is internally studentized resisuals, and "ex" is externally studentized resisuals}
}
\value{
if standardized resisuals are computed, a list of two elements will be returned
\itemize{
  \item{percentage cutoff} - {If the model assumptions are correct, z will follow a standard normal distribution
  with mean 0 and variance 1. This can be tested for whether the model follows normal distribution.}
  \item{standardized resisuals} - {all the computed standardized residuals}
}
}
\description{
res_3 computes three types of residuals from a linear regression model.
These residual types are standardized resisuals, internally studentized residuals,
and externally studentized residuals.
}
\examples{
res_3(demo_data, lm(fev~A+H+M+S, data = demo_data), r = "z")
res_3(demo_data, lm(fev~A+H+M+S, data = demo_data), r = "int")
res_3(demo_data, lm(fev~A+H+M+S, data = demo_data), r = "ex")


}
