% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.r
\name{sample_read}
\alias{sample_read}
\title{Read CSV duplicate}
\usage{
sample_read(path)
}
\arguments{
\item{path}{to file name}
}
\value{
\code{tibble}
}
\description{
Read CSV duplicate
}
\examples{
csv=system.file("extdata" , "sample_data.csv",  package = "testpackage")
sample_read(csv)
}
