% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{say_hello}
\alias{say_hello}
\title{say_hello}
\usage{
say_hello(name)
}
\arguments{
\item{name}{The person to say hello to. Default is \code{NULL}.}

\item{color}{The color of the words. Default is \code{NULL}.}
}
\description{
A function to say "hello!"
}
\author{
GM.W
}
