% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a scatterplot}
\usage{
scatterplot(data, xvar, yvar)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{xvar}{The name of the variable to be used for the x-axis.}

\item{yvar}{The name of the variable to be used for the y-axis.}
}
\value{
A ggplot2 scatterplot.
}
\description{
This function creates a scatterplot using ggplot2.
}
\examples{
data(iris)
scatterplot(iris, "Sepal.Length", "Sepal.Width")

}
