% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamCurrentFormScore.R
\name{score_team_form}
\alias{score_team_form}
\title{Team Current Form Score function}
\usage{
score_team_form(current_points, difficulty_coefficient)
}
\description{
This function assigns a score evaluating the current form 
of a team based on its first ten results of the current season
  
[INSERT MOTIVATION OF CURENT FORM FORMULA]
 
Formula: 
Team_i score = (1 - (Leader points - Team_i points) / leader points ) * last team points
}
\examples{
score_team_form()
}

