% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamQualityScore.R
\name{score_team_qual}
\alias{score_team_qual}
\title{Team Quality Score function}
\usage{
score_team_qual(team_points, leader_points, last_team_points)
}
\description{
This function assigns a score evaluating the general quality 
of a team based on its previous season relative standings.
  
[INSERT MOTIVATION OF QUALITY SCORE FORMULA]
 
Formula: 
Team_i score = (1 - (Leader points - Team_i points) / leader points ) * last team points
}
\examples{
score_team_qual()
}

