% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search2022}
\alias{search2022}
\title{Title}
\usage{
search2022(data, str)
}
\arguments{
\item{data}{A numerical matrix or table}

\item{str}{The character or number you want to search.}
}
\value{
result
}
\description{
Numeric or character positioning.
}
\examples{
testpackage::search2022(mtcars,27.3)
testpackage::search2022(mtcars,0)
}
\author{
Fei Liang
}
