% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{selectHaulIdsHiearchical}
\alias{selectHaulIdsHiearchical}
\title{resampling for hiearchical bootstrap. Strucutre is a bit different from the above to ensure that haul selection is consistent between the different tables.
simTrawlHaulsHiearchical could probably be modified to make parameters quite similar to the ones above.}
\usage{
selectHaulIdsHiearchical(hh, levels, selection, seed = NULL)
}
\arguments{
\item{hh}{Data set of hauls, formatted as the DATRAS table (could be HH, CA or HL)}

\item{levels}{vector with names of columns in hh that represents levels in the hiearchical sampling. levels[1] identifies the primary sampling unit, levels[2] the secondary, and so on.}

\item{selection}{codes for how selection is done: "R" selection with replacement (sample size: the number of available samples), "S" selection without replacement with sample size a selected as a random integer between 1 and the number of available samples (inclusive), "N": No selection or selection without replacement with sample size: the number of available samples}

\item{seed}{seed for selection, if NULL seed will not be set.}

\item{idname}{the name of a column that identifies samples}
}
\value{
a vector of haul ids, possibly repeating. Note: Does not return a data structure suitable for feeding into functions for CPUE-estimates
}
\description{
Makes hiearchical selection of haul ids for bootstrap estimation.
}
\details{
selectHaulIdsHiearchical
}
\keyword{internal}
