% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.r
\name{selsecond}
\alias{selsecond}
\title{return second column}
\usage{
selsecond(y)
}
\arguments{
\item{y}{a thing suitable to be used with dplyr}
}
\value{
a column
}
\description{
this functions return the second column
}
\details{
there are no specific details
}
\author{
Daniel Knitter
}
