% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{simCaHlSimultaniousyStratified}
\alias{simCaHlSimultaniousyStratified}
\title{simCaHlSimultaniousyStratified}
\usage{
simCaHlSimultaniousyStratified(RFA, year, quarter, dataHH, loc = NULL)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}
}
\value{
Returns simulations of the dataras-data with length information on a similar format as the data used in the functions for calculating the CPUEs.
The simulated trawl hauls are simulated stratisfied on the statistical rectangles. TODO: choose a procedure if there is only one observation in the statistical recangle, I suggest to include the closest trawl haul no matter which statistical recangles it is assosiated with..
}
\description{
Simulates trawl hauls with only length information used in the bootstrap procedure.
}
