% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sums.R
\name{simple_sums}
\alias{simple_sums}
\title{simple_sums is a load of nonsense}
\usage{
simple_sums(x)
}
\arguments{
\item{x}{is what you want to multiply by 17.4}
}
\value{
17.4 times what you put in!
}
\description{
x is multiplied by 17.4
}
\examples{
ttrrt <- simple_sums(x = 1);
df2 <- simple_sums(x = 12);
}
