% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{simTrawlHaulsCASimple}
\alias{simTrawlHaulsCASimple}
\title{simTrawlHaulsCASimple}
\usage{
simTrawlHaulsCASimple(RFA, year, quarter, data)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}
}
\value{
Returns simulations of the dataras-data with both length and age information on a similar format as the data used in the functions for calculating the CPUEs
}
\description{
Simulates trawl hauls with both length and age information used in the bootstrap procedure.
}
