% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{simTrawlHaulsCAStratified}
\alias{simTrawlHaulsCAStratified}
\title{simTrawlHaulsCAStratified}
\usage{
simTrawlHaulsCAStratified(RFA, year, quarter, data, species = "Gadus morhua")
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}
}
\value{
Returns simulations of the dataras-data with both length and age information on a similar format as the data used in the functions for calculating the CPUEs
Sample stratisfied with respect to the length observed. This gives meaning since it is assumed that the ALK is similar in the whole roundfish area.
}
\description{
Simulates trawl hauls with both length and age information used in the bootstrap procedure.
}
