% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{simTrawlHaulsHLdatras}
\alias{simTrawlHaulsHLdatras}
\title{simTrawlHaulsHLdatras}
\usage{
simTrawlHaulsHLdatras(RFA, year, quarter, data, ca_hh)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}
}
\value{
Returns simulations of the dataras-data with length information on a similar format as the data used in the functions for calculating the CPUEs
}
\description{
Simulates trawl hauls with only length information used in the bootstrap procedure.
Keep the number of trawl ahuls within each statistical rectangle fixed, and sample them with replacement from the whole RFA.
}
