% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyClass.R
\name{simulate.MyClass}
\alias{simulate.MyClass}
\title{Simulates the class object}
\usage{
\method{simulate}{MyClass}(object, ...)
}
\arguments{
\item{object}{class object}

\item{...}{other parameters}
}
\description{
Simulates the class object
}
