% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{simulateALK}
\alias{simulateALK}
\title{simulateALK}
\usage{
simulateALK(RFA, species, year, quarter, dataCA,
  bootstrapProcedure = "simple", B = 10)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALK is calculated.}

\item{quarter}{The quarter of the year which the ALK is calculated.}

\item{B}{The number of simulated ALK to be returned.}

\item{data}{The data needed for calculating the ALK.}
}
\value{
Returns a list with simulated ALK given data, species, time, RFA and simulation procedure
}
\description{
Simulate a given number of ALK-key-matrix for a given year, quarter, species and RFA.
}
