% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites-data.R
\docType{data}
\name{sites}
\alias{sites}
\title{Abundance of five species on 12 sites}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 5 columns.
}
\source{
Random values
}
\usage{
sites
}
\description{
A dataset containing randomly generated abundance of five species
on 12 sites.
}
\keyword{datasets}
