% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Return the square of a number.}
\usage{
square(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The square of \code{x}.
}
\description{
Return the square of a number.
}
\examples{
#' square(2)
}
