% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square.R
\name{Square_plus_1}
\alias{Square_plus_1}
\title{Test Function: Square_plus_1}
\usage{
Square_plus_1(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a vector.
}
\description{
Test Function: Square_plus_1
}
\examples{
x <- 1:3
Square_plus_1(x)
}
