% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{squared}
\alias{squared}
\title{Get square of a real number
This function take a real number and square the value.}
\usage{
squared(x)
}
\arguments{
\item{x}{}
}
\value{
squared number
}
\description{
Get square of a real number
This function take a real number and square the value.
}
\examples{
squared(12)
}
