% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTplot.R
\name{SSTplot}
\alias{SSTplot}
\title{SST off the West Coast of India}
\usage{
SSTplot(date = "last")
}
\arguments{
\item{date}{A character string in "2014-01-30" format. The default is "last" which is the latest values.}
}
\value{
A plot of SST.
}
\description{
This little function plots the SST off the western coast of India from remote sensing data.
}
\details{
These data are from the \href{https://podaac.jpl.nasa.gov/dataset/UKMO-L4HRfnd-GLOB-OSTIA}{Global SST & Sea Ice Analysis},
L4 OSTIA, UK Met Office,
Global, 0.05°, Daily,2013-present dataset available on ERDDAP at
\href{https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplUKMO_OSTIAv20}{ERDDAP}.

The default is to show the
latest SST values but you can enter another date.  The earliest year in this data
set is 2013.
}
\examples{
SSTplot("2017-09-09")
}
