% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_for_package_content.R
\name{stacked_bar}
\alias{stacked_bar}
\title{stacked_bar}
\usage{
stacked_bar(data, x_axis, y_axis, title, legend_title, x_label, y_label, color)
}
\arguments{
\item{data}{Data frame containing the variables to plot}

\item{x_axis}{Column name for x-axis values}

\item{y_axis}{Column name for y-axis values}

\item{title}{Plot title}

\item{legend_title}{Legend title}

\item{x_label}{X-axis label}

\item{y_label}{Y-axis label}

\item{color}{Vector with hex code colors}
}
\value{
A ggplot object
}
\description{
Create a stacked bar plot using ggplot2.
}
\examples{
# stacked_bar(df, 'category', 'value', 'Title', 'Legend', 'X', 'Y', c('#123456','#654321'))
}
