% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strToVector.R
\name{strToVector}
\alias{strToVector}
\title{Convert string to vector with only the string}
\usage{
strToVector(string)
}
\arguments{
\item{string}{A character string to be converted to a vector}
}
\value{
A character vector containing only the input string
}
\description{
This function takes a string as input and returns a vector with only the string in it.
}
\keyword{internal}
