% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyClass.R
\name{summary.MyClass}
\alias{summary.MyClass}
\title{Summarizes the class object}
\usage{
\method{summary}{MyClass}(object, ...)
}
\arguments{
\item{object}{class object}

\item{...}{other parameters}
}
\description{
Summarizes the class object
}
