% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotaSamplingCheck.R
\name{tabulateMissingAgeSamples}
\alias{tabulateMissingAgeSamples}
\title{Tabulates the number of missing age samples in each length group and roundfish area.}
\usage{
tabulateMissingAgeSamples(hh = HH, ca = CA, hl = HL,
  species = "Gadus morhua")
}
\arguments{
\item{hh}{HH table for the survey}

\item{ca}{CA table for the survey}

\item{hl}{HL table for the survey}

\item{scientific}{name of species}
}
\value{
data frame with columns: LngtCM RFA count.age count.length missing
}
\description{
Tabulates the number of missing age samples in each length group and roundfish area.
}
