\name{test1}
\alias{test1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 test 1
}

\usage{
 test1(par1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par1}{
  A number .....
}
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
   test1(20)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
