% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunction.R
\name{testFunction}
\alias{testFunction}
\title{Title of a test function}
\usage{
testFunction(x)
}
\arguments{
\item{x}{item to be printed}
}
\value{
Prints two copies of x
}
\description{
This function prints two copies of the input
}
