% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackage-package.R
\docType{package}
\name{testPackage-package}
\alias{testPackage}
\alias{testPackage-package}
\title{testPackage: A remplir}
\description{
Utilities for automating the construction and development of an R package for submission to CRAN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecamenen/testPackage}
  \item Report bugs at \url{https://github.com/ecamenen/testPackage/issues}
}

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

}
\keyword{internal}
