% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{testpackage.mean}
\alias{testpackage.mean}
\title{Mean Function}
\usage{
testpackage.mean(arg0 = NA)
}
\arguments{
\item{arg0}{Number vector to calculate mean.}
}
\description{
This function calculates the mean of a number vector.
}
\examples{
vector <- c(5,3,8,4,5,6)
testpackage.mean(vector)
}
\keyword{average}
\keyword{mean}

