% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: A package to demonstrate R package utility for educators}
\description{
This package contains a raw data file, an RMarkdown template, and a LearnR tutorial.  
    This demonstrates that packages are a useful way for educators to deliver materials and teaching content to students.
}
\author{
\strong{Maintainer}: Desirée De Leon \email{desiree@rstudio.com} (https://orcid.org/0000-0003-2217-9067)

}
\keyword{internal}
