% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{testpackage.sd}
\alias{testpackage.sd}
\title{SD Function}
\usage{
testpackage.sd(arg0 = NA)
}
\arguments{
\item{arg0}{number vector.}
}
\description{
This function calculates the standard deviation of a number vector.
}
\examples{
vector <- c(5,3,8,4,5,6)
testpackage.sd(vector)
}
\keyword{sd}

