% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_IMD.r
\name{theme_IMD}
\alias{theme_IMD}
\title{theme_IMD: custom ggplot2 theme for plotting}
\usage{
theme_IMD()
}
\value{
This function must be used in conjunction with a ggplot object, and will return a ggplot object with the custom theme.
}
\description{
This is a custom ggplot2 theme that removes the default panel grids
from ggplot2 figures, and makes the axes and tick marks grey instead of black.
}
\examples{
example_dat <- data.frame(Latin_Name = c("Carex limosa", "Arethusa bulbosa",
                                         "Malaxis unifolia", "Calopogon tuberosus"),
                          cover = c(10, 40, 10, 50),
                          stems = c(50, 20, 10, 10))
library(ggplot2)
p <- ggplot(data = example_dat, aes(x = cover, y = stems)) +
     geom_point() +
     theme_IMD()
p

}
