% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage.R
\name{tnl.sim}
\alias{tnl.sim}
\title{The Tnl simulation function.}
\usage{
tnl.sim(k, n, l, trial = 1e+05)
}
\arguments{
\item{k}{Vector of quantiles.}

\item{n}{Sample size.}

\item{l}{The class parameter of Tnl.}

\item{trial}{The number of trials.}
}
\value{
list of simulated probability and distribution function.
}
\description{
Perform simulated tnl.
}
\examples{
\dontrun{You need to select all of \code{k}, \code{n} and \code{l}}
\dontrun{\code{trial}=100000 by default}
tnl.sim(k=2,n=7,l=1)
\dontrun{
$prob
[1] 0.15476

$tt
[1] 0.23158
}
}
