% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TokenizeString.R
\name{TokenizeString}
\alias{TokenizeString}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
TokenizeString(string)
}
\arguments{
\item{string}{A single input string such as "This is a cool function!"}
}
\value{
A vector contain all valid tokens in the original input string
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}

