% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial.R
\name{trivial}
\alias{trivial}
\title{trivial function}
\usage{
trivial(x, y)
}
\arguments{
\item{x, y}{two variables to be summed}
}
\value{
the sum of x + y
}
\description{
trivial function
}
\examples{
trivial(x,y)
}
