% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ufo}
\alias{ufo}
\title{UFO sighting data from 1969 to 2068}
\format{A data frame with 102240 rows and 8 variables:
\describe{
  \item{date}{date of sighting}
  \item{date_time}{date and time of sighting, if time available}
  \item{city}{city}
  \item{state}{state}
  \item{shape}{shape}
  \item{duration}{duration of encounter}
  \item{summary}{written description}
  \item{posted}{when added to database}
}}
\source{
\url{http://www.nuforc.org/}
}
\usage{
ufo
}
\description{
A dataset containing data on historical UFO sightings
}
\keyword{datasets}
