% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tables.R
\name{ufo_data}
\alias{ufo_data}
\title{Import the ufo table}
\usage{
ufo_data(include_descrip = FALSE)
}
\arguments{
\item{include_descrip}{Whether to include the event description,
Default is FALSE}
}
\value{
Returns the ufo sightings table.  See \link{ufo}
}
\description{
Import the ufo table
}
\examples{
ufo = ufo_data()

}
