% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate that the parameters meet the required schema}
\usage{
validate_parameters(params, pschema = "pca_projection_schema.json")
}
\arguments{
\item{lparams}{The list of parameters to be validated}
}
\description{
Validate that the parameters meet the required schema
}
