% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waitApiCooldown.R
\name{waitApiCooldown}
\alias{waitApiCooldown}
\title{Wait for a specified number of seconds}
\usage{
waitApiCooldown(seconds = 10)
}
\arguments{
\item{seconds}{The number of seconds to wait (default = 60)}
}
\description{
This function waits for a specified number of seconds before continuing.
}
\keyword{time}
\keyword{wait}
