% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Adds two numbers together}
\usage{
add(a, b)
}
\arguments{
\item{a}{Numeric variable}

\item{b}{Numeric variable}
}
\value{
Numeric sum of variables
}
\description{
Adds two numbers together
}
