% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(a, b)
}
\arguments{
\item{a}{numeric vector of length one}

\item{b}{numeric vector of length one}
}
\value{
numeric vector of length one
}
\description{
Add two numbers
}
\examples{
add_numbers(4.5, 7)
}
