% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUEnorthSea}
\alias{calcmCPUEnorthSea}
\title{calcmCPUEnorthSea}
\usage{
calcmCPUEnorthSea(rfa, species, year, quarter, dat, bootstrapProcedure, B,
  mCPUE, column)
}
\arguments{
\item{rfa}{information about size of RFA's.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{column}{column in the mCPUE matrix to calculate mCPUE}

\item{data}{the datras-data needed.}

\item{percentOfAreaRepresentative}{the percentage of the statical recangle within sea depth intervall}
}
\value{
Returns the mCPUE per length class in the given statistical rectangle
}
\description{
Internal function for calculating mCPUE in whole North sea.
}
