% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{calcMCPUEwholeNorthSea}
\alias{calcMCPUEwholeNorthSea}
\title{calcCPUEwholeNorthSea}
\usage{
calcMCPUEwholeNorthSea(species, year, quarter, dat,
  bootstrapProcedure = "simple", B = 10, removeProportionsOfCA = 0,
  removeProportionsOfHL = 0, procedure = "datras")
}
\value{
Returns the mCPUE per age class in the whole North Sea
}
