% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean.R
\name{compute_mean}
\alias{compute_mean}
\title{Compute the mean of the column in a data.frame}
\usage{
compute_mean(column_name, path)
}
\arguments{
\item{column_name}{character: Column name for which the mean is to be
computed}

\item{path}{character: Path to the data file}
}
\value{
numeric
}
\description{
Compute the mean of the column in a data.frame
}
