% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{CPUEnorthSea}
\alias{CPUEnorthSea}
\title{calcCPUEwholeNorthSea}
\usage{
CPUEnorthSea(species, year, quarter, dat, bootstrapProcedure = "datras",
  B = 10, ALKprocedure = "", doBootstrap = TRUE, useFisher = FALSE,
  removeFirst = FALSE, lengthDivision = NULL, removeProcedure = NULL)
}
\value{
Returns the mCPUE per age class in the whole North Sea
}
\description{
calcCPUEwholeNorthSea
}
