% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Sample Data for Demostration}
\format{
An object of class \code{data.frame} with 654 rows and 6 columns.
}
\usage{
data(demo_data)
}
\description{
A common measurement of lung function is the forced expiratory
volume FEV, which measures how much air a person can blow out of
the lungs in a short period of time. A higher FEV is usually
associated with better respiratory function. A group of clinicians is
interested in examining the factors which affect FEV1. Data on 654
children are available for the following covariates: AGE (A, years),HEIGHT (H, inches),
MALE (M=0 for female, 1 for male) SMOKE(Si=0 for non-smoker and 1 for smoker).
Each row corresponds to a single clinic visit.
}
\references{
Module K, BIOSTAT650, Dr. Xu Shi ,University of Michigan, School of Public Health
}
\keyword{datasets}
