% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBSeqTest.R
\name{EBSeqTest}
\alias{EBSeqTest}
\title{iterative pruning empirical bayesian}
\usage{
EBSeqTest(data, conditions, uc, iLabel = 1, sizefactor = 1,
  iter = 50, alpha = 0.4, beta = 0, step1 = 1e-06, step2 = 0.01,
  thre = log(2), sthre = 0.001, filter = 10, stopthre = 0.001)
}
\arguments{
\item{data}{single cell expression matrix, row as genes column as cells (normalized counts required)}

\item{conditions}{partition of cells}

\item{uc}{number of unceratin relations between means of subtypes for each gene level}

\item{iLabel}{label for isoform, indicating which beta are shared}

\item{iter}{maximum iteration step of EM}

\item{alpha}{start point of hyper parameter alpha}

\item{beta}{start point of hyper parameter beta}

\item{step1}{stepsize for gradient ascent of alpha}

\item{step2}{stepsize for gradietn ascent of beta}

\item{thre}{threshold for determining whether a relation is sure or uncertain}

\item{sthre}{shrinkage threshold for iterative pruning space of DE patterns}

\item{filter}{filterthreshold for low expression gene for DE analysis}

\item{stopthre}{stopping threshold for EM}
}
\value{
a list containing considered DE patterns and their posterior probability, values for alpha and beta
}
\description{
iterative pruning empirical bayesian
}
