% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{My second function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{an n x 1 vector of observations}

\item{X}{an n x k vector of observations}
}
\value{
beta_hat An estimate of beta in y = beta X + epsilon
}
\description{
My second function
}
