% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Simple Linear Regression}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{is a n x 1 vector of 1 dependent variables.}

\item{X}{is a n x k matrix of k independent variables}
}
\value{
beta_hat, a k x 1 vector of linear regression coefficients
}
\description{
Simple Linear Regression
}
