% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate Beta for Linear Regression}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numeric vector of n observations.}

\item{X}{A numeric matrix of n x k predictor variables.}
}
\value{
A numeric vector of estimated beta coefficients.
}
\description{
Estimate Beta for Linear Regression
}
\examples{
y <- c(2, 3, 5, 7)
X <- matrix(c(1, 1, 1, 1, 1, 2, 3, 4), ncol = 2)
estimate_beta(y, X)
}
