% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{My estimate_beta() function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector.}

\item{X}{A n x k numerical matrix.}
}
\value{
Vector beta from the regression y = beta*x + e.
}
\description{
My estimate_beta() function
}
