% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{My first function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A n dimensional numerical vector.}

\item{X}{A numerical n x K matrix.}
}
\value{
A K dimensional numerical vector of coefficients from the linear regression of y on X.
}
\description{
My first function
}
\examples{
myfunc(3, 5)
}
