% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{My first function}
\usage{
estimate_beta(y, x)
}
\arguments{
\item{y}{A n dimensional numerical matrix.}

\item{x}{A numerical n x K matrix.}
}
\value{
A K dimensional numerical matrix of coefficients from the linear regression of y on x.
}
\description{
My first function
}
\examples{
myfunc(3, 5)
}
