% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{OLS Beta estimation}
\usage{
estimate_beta(X, y)
}
\arguments{
\item{X}{NxK Matrix.}

\item{y}{Nx1 numerical vector.}
}
\value{
b kx1 numerical vector containg OLS coefficients.
}
\description{
OLS Beta estimation
}
\examples{
estimate_beta(c(1,0,4,5,4), c(9,0,9,2,1))
}
