% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Gives the OLS estimator}
\usage{
estimate_beta(X, y)
}
\arguments{
\item{X}{n by k matrix}

\item{y}{n by 1 matrix}
}
\value{
k by 1 vector of coefficients
}
\description{
Gives the OLS estimator
}
\examples{
estimate_beta(X, y)
}
