% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{A function to estimate beta for a linear regression.}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A vector of n observations.}

\item{X}{A nxk matrix of k variables.}
}
\value{
A vector of k variables.
}
\description{
A function to estimate beta for a linear regression.
}
\examples{
estimate_beta(c(1, 2, 3), matrix(data = 1:6, ncol = 2))
}
