% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{My function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector.}

\item{X}{Also a numerical vector.}
}
\value{
A numerical
}
\description{
My function
}
\examples{
estimate_beta(3, 5)

}
