% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate the OLS coefficients}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector of size n*1.}

\item{b}{A numerical matrix of size n*k.}
}
\value{
A numerical vector of size k*1.
}
\description{
Estimate the OLS coefficients
}
\examples{
X <- matrix(c(3, 2, 5, 1), nrow = 2)
y <- c(5, 1)
estimate_beta(y, X)
}
