% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{My second function}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector.}

\item{X}{A numerical matrix.}
}
\value{
A numerical vector of (x'x)^-1*(x'y).
}
\description{
My second function
}
\examples{
estimate_beta(c(3, 5), matrix(c(10, 8, 5, 12), ncol = 2, byrow = TRUE))
}
