% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate Linear Regression Coefficient}
\usage{
estimate_beta(x, y)
}
\arguments{
\item{x}{An NxK matrix of K variables}

\item{y}{A Nx1 vector of observations}
}
\value{
beta A Kx1 vector of linear regression coefficients: beta = (X'X)^-1 (X'Y)
}
\description{
Estimate Linear Regression Coefficient
}
