% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate Regression Coefficients using OLS}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numeric vector of response variable values (n observations).}

\item{X}{A numeric matrix of predictor variables (n x k).}
}
\value{
A numeric vector of estimated coefficients \eqn{\hat{\beta}}.
}
\description{
This function computes the ordinary least squares (OLS) estimator for a linear regression model.
}
\examples{
y <- iris[, "Sepal.Length"] |> as.matrix()
X <- iris[, c("Sepal.Width", "Petal.Length", "Petal.Width")] |> as.matrix()
estimate_beta(y, X)
}
