% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Function to estimate beta}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A vector of n observations where n>1.}

\item{X}{An n*k matrix of k variables where n>=1.}
}
\value{
Estimate beta for linear regression.
}
\description{
Function to estimate beta
}
\examples{
y <- iris[, "Sepal.Length"] |> as.matrix()
X <- iris[, c("Sepal.Width", "Petal.Length", "Petal.Width")] |> as.matrix()
estimate_beta(y, X)
}
