% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.r
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimate Beta for Linear Regression}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{A numerical vector of response values.}

\item{X}{A numerical matrix of predictor values.}
}
\value{
A numerical vector of estimated beta coefficients.
}
\description{
Estimate Beta for Linear Regression
}
\examples{
y <- c(1, 2, 3, 4, 5)
X <- matrix(c(1, 1, 1, 1, 1, 1, 2, 3, 4, 5), ncol = 2)
estimate_beta(y, X)
}
