% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.r
\name{estimate_beta}
\alias{estimate_beta}
\title{estimate_beta}
\usage{
estimate_beta(y, X)
}
\arguments{
\item{y}{: a n by 1 vector}

\item{X}{: a n by k matrix.}
}
\value{
A ols estimate of beta
}
\description{
estimate_beta
}
\examples{
estimate_beta(y,X)
}
