% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates accident maps}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value, which corresponds to the State Code.}

\item{year}{A character string or an integer, which determines the year.}
}
\description{
Creates a map for a state and a year to show where the corresponding
accicdents occured.
}
\examples{
\dontrun{
fars_map_state(49,2015)
}

}
