% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plot the geographic locations of accidents in a state during a certain year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a character or integer vector of length 1 representing the state number intended}

\item{year}{a character or integer vector of length 1 representing the year desired}
}
\value{
This function returns NULL and plots the locations of accidents on map in a state
}
\description{
Takes \code{state.num} and \code{year} as arguments and plots the locations of the accidents in a state
during a year.
}
\examples{
fars_map_state(1,2013)
fars_map_state("1", "2013")

}
