% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show a plot a map with the year's accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer with the state code}

\item{year}{An integer with the year}
}
\value{
A plot of the state with points on the accident
}
\description{
this plot a map for the state provided showing the accidents of the year geolocalized
}
\details{
An error is raised if the state code is not valid

An error is raised if doesnt exists accidentes data for the state plotted

Use the maps and graphics packages
}
\examples{
fars_map_state(42,2015)

}
