% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map Fatality Analysis Reporting System Data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{\code{integer} number that specifies state number}

\item{years}{\code{vector} of \code{numerical} value that converted to the year, used to specify filename}
}
\value{
\code{fars_map_state} plots a map of the accidients based on year
and state specified. If there is no accident, a prompt message returns. 
If an invalid state number is provided, an error will return.
}
\description{
\code{fars_map_state} plots the Fatality Analysis Reporting System data
based on year and state specified
}
\examples{
fars_map_state(2, 2016)

}
