% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{The Map of Records in a State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character string or a integer giving the state number.}

\item{year}{An integer specifying a year.}
}
\value{
A map featuring the boundary of the state based on the records,
   the state, and also the recoreds as dots inside the state.
}
\description{
This is a function that draws the map of the given state, and the
   records are shown on the map as dots according to their locations.
}
\details{
1. the function will stop with an error. 2. If the input state.num does not match any
   record, the function will stop with the error "invalid STATE number". 3. If no records match
   the given year and state, the function will stop with the message "no accidents to plot".
}
\examples{
\dontrun{
 fars_map_state(5, 2014)
 fars_map_state(3, 2013)
}

}
