% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageproject.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character vector.}
}
\value{
If the file named exists, the function returns a dataframe. If the
file named does not exist, then an error message is returned.
}
\description{
This function reads in data from a file.  If the file named does not exist,
an error message is shown.
}
\examples{
\dontrun{fars_read("filename")}
}
