% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the text the function will print}
}
\value{
This function returns a phrase to print, with or without an 
   exclamation point added. As a side effect, this function also prints out
   the phrase.
}
\description{
This is a simple function that checks to see if a file exists. You can 
customize the text to print (using the \code{to_print} argument) and add
an exclamation point (\code{excited = TRUE}).
}
\examples{
fars_read("test")

}
