% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{pre_reading checking for existence of a .csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a path to a file, a connection, or literal data}
}
\value{
This function returns a \code{tbl_df} representing the contents of \code{filename}
}
\description{
Checks for existence of a .csv file. If it exists, \code{fars_read} returns the contents as
an object of class "tbl_df". If the file does not exist, \code{fars_read} returns an error.
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
fars_read("accident_2014.csv")
fars_read("accident_2015.csv")
##long paths
read_csv("https://github.com/tidyverse/readr/raw/master/inst/extdata/mtcars.csv")
 }

}
