% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load a FARS data file in R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The datafile filename}
}
\value{
A data frame. As a side effect, this function also prints out the first lines of the data frame loaded
}
\description{
This simple function read a FARS datafile located in the default directory
}
\details{
An error is raised if the filename doesnt exist

Uses the dplyr and readr packages
}
\examples{
fars_read("accident_2013.csv.bz2")

}
