% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{\code{string} of characters that contain filename of the csv file that contains the data}
}
\value{
\code{tibble} \code{data.frame} of the data read from csv file if such filepath exists;
it returns an error message returns if specified filename does not exist in the path
}
\description{
\code{fars_read} reads the Fatality Analysis Reporting System
}
\examples{
\dontrun{fars_read('daily_SPEC_2014.csv.bz2')}

}
