% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv.bz2 into data.frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the compressed
file.}
}
\value{
This function returns a data.frame containing the data
   in the .csv file.
}
\description{
This is a function that reads the .csv file in a compressed file into a
   data.frame in R.
}
\details{
The directory of the file should be the same as the working
   directory. Otherwise the function will stop with an error.
}
\examples{
\dontrun{
 fars_read("accident_2013.csv")
}


}
