% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A List of strings or numbers, which correspond to the years from
which to extract data.}
}
\value{
A tibble including entries of month and year read from the different
  csv files of the different years provided in \code{years}, or \code{NULL}
  and a warning if the input is not valid.
}
\description{
An auxiliary function used by \code{fars_summarize_years} to extract years
and months from FARS data.
}
\examples{
data <- fars_read_years(c(2014, 2015))

}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}

\code{\link{fars_summarize_years}}
}
