% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageproject.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integer(s).}
}
\value{
A dataframe with two variables (Month and year) or an error message.
}
\description{
This function takes in an integer vector and using the \code{\link{make_filename}}
function creates file names.  Then using the \code{\link{fars_read}} function
it searches if those files exist. If the files exist and are the correct
format then a dataframe is returned, otherwise a warning is returned. The
returned dataframe has two variables (Month and year).  If multiple years
are provided, multiple dataframes are returned.
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(2013:2015)}

}
