% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract \code{YEAR} and \code{MONTH} columns from US yearly accidents data set.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a character or integer vector containing the years}
}
\value{
This function returns a \code{list} with each element of the list containing the \code{YEAR}
 and \code{MONTH} columns of each file. It returns NULL for a file it is not found.
}
\description{
Takes a character or integer vector of years as input, reads the file of each year, and then
extracts the \code{YEAR} and \code{MONTH} columns from each file. Throws a warning and returns NULL.
for the unfound file.
}
\examples{
\donttest{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c("2013", "2014", "2015"))
}

}
