% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Filter the FARS data by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list with the years to filter}
}
\value{
A data frame with two columns YEAR and MONTH. As a side effect, this function also prints out the first lines of the data frame loaded
}
\description{
Filter the FARS data for the years passed as a list and return a dataframe with two columns: MONTH and YEAR
}
\details{
An error is raised if doesnt exist data for the year passed

Uses the dplyr and magrittr packages
}
\examples{
fars_read_years(c(2013,2014,2015))

}
