% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fatality Analysis Reporting System File based on years specified}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{\code{vector} of \code{numerical} values of specified years}
}
\value{
\code{fars_read_years} will take in multiple years, and create standardized filenames,
and then search for and read the filenames and return \code{tibble} \code{data.frame}. 
If such filename with the specified year does not exist, an error message will return
}
\description{
\code{fars_read_years} takes in multiple years, and returns the
Fatality Analysis Reporting System Data based on the years specified
}
\examples{
fars_read_years(2010:2017)

}
