% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Fetch 'Year' and 'Month' of the Records}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector that specifying the years of interest.}
}
\value{
A list. Each element is a data.frame containing the corresponding
   'year' and 'MONTH' columns.
}
\description{
This is a function that fetches the 'year' and 'MONTH' columns of the
   records of regarding fatal injuries, in the specific given years.
}
\details{
If a file for a year does not exist in the working directory,
   the return value will be NULL with a warning message.
}
\examples{
\dontrun{
 fars_read_years(c(2013,2014,2015))
 fars_read_years(c(2013,2016))
}

}
