% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageproject.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integer vector}
}
\value{
A table of fatalities summarized by month by year.
}
\description{
This function takes in a list of years and outputs a summarized table of
fatalities by month by year.  If the data files do not exist or are in the
incorrect format an error message will appear.
}
\examples{
\dontrun{fars_summarize_years(2013)}

}
