% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents for per month for each year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{character or integer vector containing the years}
}
\value{
This function returns a \code{tbl_df} representing the summary of number of accients per month for each year.
}
\description{
Takes a character or integer vector of years and returns a dataframe. The dataframe's first column
is \code{MONTH} and the rest represent the year with each observation representing the month
and number of accidents for each year. Shows a warning for each invalid year.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
fars_summarize_years(c("2013","2014","2015"))
}

}
