% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the rows of FARS data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list with the years to filter}
}
\value{
A data frame with the data summarize by month
}
\description{
Summarize the FARS data by month for the years passed as a list and return a dataframe with one row by month and one column for each year
}
\details{
Uses the dplyr, tidyr and magrittr packages
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
