% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Fatality Analysis Reporting System Data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{\code{vector} of \code{numerical} values of specified years}
}
\value{
\code{fars_summarize_years} returns summarized result in \code{data.frame} of Fatality Analysis
Reporting System Data based on year specified
}
\description{
\code{fars_summarize_years} reads in Fatality Analysis Reporting
  System data files based on the years specified and summarized the
  result based on the year and month.
}
\examples{
\dontrun{fars_summarize_years(2015:2016)}

}
