% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the counts of fatal injuries}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector that specifying the years of interest.}
}
\value{
A data.frame with 12 rows and \code{length(years)} columns.
    Each row represents a month, and each column represents a year.
}
\description{
This is a function that Summarize the counts of fatal injuries
   for each month, given specific years.
}
\details{
If a file for a year does not exist in the working directory,
   the returned data.frame will not contain the corresponding column,
   with a warning message.
}
\examples{
\dontrun{
 fars_summarize_years(c(2013,2014))
 fars_summarize_years(c(2013,2015))
}

}
