% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Helper function to load data into memory}
\usage{
get_data(path)
}
\arguments{
\item{path}{character: Path to .parquet file}
}
\value{
data.frame
}
\description{
Helper function to load data into memory
}
