% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_img.R
\name{get_img}
\alias{get_img}
\title{Create a graphical object (GROB) from png or jpg}
\usage{
get_img(
  filename,
  FUN = png::readPNG,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = NULL,
  height = NULL,
  just = "centre",
  vjust = NULL,
  hjust = NULL,
  gp = grid::gpar()
)
}
\arguments{
\item{filename}{Filepath to png or jpg file.}

\item{FUN}{Import function. Switches internally between \code{png::readPNG} and \code{jpeg::readJPEG}}

\item{x}{A numeric vector or unit object specifying x-location.}

\item{y}{A numeric vector or unit object specifying y-location.}

\item{width}{A numeric vector or unit object specifying width.}

\item{height}{A numeric vector or unit object specifying height.}

\item{just}{The justification of the rectangle
    relative to its (x, y) location.  If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment.
  }

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{gp}{An object of class \code{"gpar"}, typically the output
    from a call to the function \code{\link[grid]{gpar}}.  This is basically
    a list of graphical parameter settings.}
}
\value{
A \code{GROB} that can be added to a ggplot.
}
\description{
Create a graphical object (GROB) from png or jpg
}
