% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetNGrams.R
\name{GetNGrams}
\alias{GetNGrams}
\title{A function to get n-grams from a tokenized sentence.}
\usage{
GetNGrams(words, n)
}
\arguments{
\item{words}{A vector of strings that represents a tokenized document}

\item{n}{The size of the n-gram window desired}
}
\description{
A function to get n-grams from a tokenized sentence.
}

