\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello(inval)
}
\description{
Prints 'Hello, world!'.
}
\arguments{
  \item{inval}{a connection or a character string giving the
    name of the file to load.}
}
\details{
This section is saved for generic details.
}
\examples{
hello(2)
}
\author{
Chris Rook\cr Email: \email{cr883296@gmail.com}\cr LinkedIn: \url{https://www.linkedin.com/in/christopher-rook-princeton/}\cr Channel: \url{https://www.youtube.com/@ChrisR-b7s}\cr
}
