% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello and multiply a number by 2.}
\usage{
hello(inval)
}
\arguments{
\item{inval}{A number.}
}
\value{
A number.
}
\description{
Say hello and multiply a number by 2.
}
\examples{
hello(5)
hello(10)
}
