% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say Hello}
\usage{
hello(n = 1)
}
\arguments{
\item{n}{integer, number of times to repeat hello (default is 1)}
}
\description{
Say hello a variable number of times
}
\examples{
hello()
hello(n = 4)

}
