% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{investigateRemoval}
\alias{investigateRemoval}
\title{investigateRemovalAfterMeeting}
\usage{
investigateRemoval(species, year, quarter, dat,
  bootstrapProcedure = "stratifiedHLandCA", B,
  ALKprocedure = "haulBased", doNotRemoveAbove = 9999, lengthDivision,
  samplesWithinEachIntervall = 1)
}
\arguments{
\item{species}{The species of interest.}

\item{year}{The year species of interest.}

\item{quarter}{The quarter species of interest.}

\item{dat}{Data object}

\item{bootstrapProcedure}{Bootstrap procedure used.}

\item{B}{Samples in the bootstrap}

\item{ALKprocedure}{ALK procedure used for calculating CPUE}

\item{doNotRemoveAbove}{Do not remove fish that is larger than this (in cm).}

\item{lengthDivision}{Take one otholit from each intervall (in cm).}

\item{samplesWithinEachIntervall}{Number of samples taken within each length intervall.}
}
\value{
Returns a list with summary about changes in estimates by removal of data.
}
\description{
Remove a proportion of the data in a given way. Return a list regarding changes of summery statistics.
}
