% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{investigateRemoval}
\alias{investigateRemoval}
\title{removeDataDetailed}
\usage{
investigateRemoval(RFA, species, year, quarter, dat, bootstrapProcedure, B,
  ALKprocedure, removeProcedure, prop, outerBootstrapN, whatToInvestigate)
}
\arguments{
\item{RFA}{The roundfish area of interest.}

\item{removeProcedure}{Removal procedure, given as an integer. removeProcedure==1 means that data is removed at random.}

\item{prop}{proportion of data to remove.}

\item{datDetailed}{The quarter of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
