% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Creates a data file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Number or string. This value will be added to the end of the
file name.}
}
\value{
If the input is a number or a string, then the output will be a
  character string containg the year at the end. Otherwise it will throw an
  error.
}
\description{
\code{make_filename} creates a filename which finishes with the stated year.
}
\examples{
filename1 <- make_filename(2014)
filename2 <- make_filename("2015")

}
