% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageproject.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer vector.}
}
\value{
A character vector to be used as a file name.
}
\description{
This function creates a name for a file given a year.
}
\examples{
\dontrun{make_filename(2009)}

}
