% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Generate an accidents file from a year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a vector of length 1 with the year desired}
}
\value{
This function returns a character vector representing the file name of the accidents at the \code{year}
}
\description{
Generates a character vector of length 1 with representing the file name storing the american
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes of the \code{year}
inputted
}
\examples{
make_filename(2013)
make_filename("2013")

}
