% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create the FARS filename for a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An year in numeric format}
}
\value{
A string with the FARS filename
}
\description{
This simple function create the filename of a FARS file for the year passed
}
\examples{
make_filename(2015)

}
