% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{\code{numerical} value that converted to the year, used to specify filename}
}
\value{
\code{make_file} returns \code{string} of characters that contain
csv filename based on the year specified
}
\description{
\code{make_file} creates a standardized filename based on the year specified
}
\examples{
\dontrun{make_filename(2017)}

}
