% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Standardized Filename Based on a Specific Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer specifying a year.}
}
\value{
This function returns a character string with the format
   "accident_%(year).csv.bz2".
}
\description{
This is function that makes a standardized filename based on a
   specific year. For example, with input 2021, the output will
   be a character string "accident_2021.csv.bz2".
}
\examples{
make_filename(1990)

}
