% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_waffle.R
\name{MakeWaffle}
\alias{MakeWaffle}
\title{Make a Waffle Plot}
\usage{
MakeWaffle(vec, title, colors)
}
\arguments{
\item{vec}{a numeric vector of strictly non-negative values}

\item{title}{the title of the plot. Defaults to the name of \code{vec}.}

\item{colors}{a character vector of colors the same length as \code{vec}.
Defaults to colors from \code{\link[grDevices]{rainbow}}.}
}
\value{
Nothing. A waffle plot is displayed.
}
\description{
Make a Waffle Plot
}
\examples{
  MakeWaffle(vec = runif(5), title = "Waffle plot of random proportions")
}
