\name{my_pkg_data}
\alias{my_pkg_data}
\docType{data}
\title{
      dataset/random number
}
\description{
            1000 random sampling
}
\usage{data("my_pkg_data")}
\format{
  The format is:
 int [1:1000] 752 208 31 140 359 764 3 420 485 564 ...
}
\details{
          1000 random sampling
}
\source{
        1000 random sampling
}
\references{
            possibly secondary sources and usages
}
\examples{
data(my_pkg_data)
## maybe str(my_pkg_data) ; plot(my_pkg_data) ...
}
\keyword{datasets}
