% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{GGPlot Theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object
}
\description{
GGPlot Theme
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(x = wt, y = mpg)) + ggplot2::geom_point() + my_theme()
}
