% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.r
\name{my_theme}
\alias{my_theme}
\title{My third function}
\usage{
my_theme()
}
\value{
A ggplot2 theme object.
}
\description{
My third function
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  my_theme()
}
