% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 Theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object with custom styles.
}
\description{
Custom ggplot2 Theme
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
ggplot2::geom_point() + my_theme()
}
