% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{Apply a custom theme to a ggplot object}
\arguments{
\item{plot}{A ggplot object to which the theme will be applied.}
}
\value{
A ggplot object with the custom theme applied.
}
\description{
This function applies a minimal theme to a ggplot object and customizes the
plot title and axis titles.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
my_theme(p)
}
