% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom Theme.r
\name{my_theme}
\alias{my_theme}
\title{My Custom ggplot2 Theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object.
}
\description{
This function returns a ggplot2 theme that modifies a standard base theme
with at least three customizations.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  my_theme()
}
