% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{A function that creates my favorite ggplot2() theme.}
\usage{
my_theme()
}
\value{
My favorite ggplot2() theme.
}
\description{
A function that creates my favorite ggplot2() theme.
}
\examples{
library(ggplot2)
p <-
  iris |>
  ggplot(aes(
    x = Sepal.Length,
    y = Sepal.Width
  )) +
  geom_point() +
my_theme()
p
}
