% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{A function that returns my customized ggplot2 theme.}
\usage{
my_theme()
}
\value{
The customized ggplot2 theme.
}
\description{
A function that returns my customized ggplot2 theme.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + my_theme()
}
