% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{A function to estimate beta for a linear regression.}
\usage{
my_theme()
}
\value{
A vector of k variables.
}
\description{
A function to estimate beta for a linear regression.
}
\examples{
my_theme()
}
