% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.r
\name{my_theme}
\alias{my_theme}
\title{Apply a ggplot2 theme to a plot}
\usage{
my_theme(plt)
}
\arguments{
\item{plt}{A ggplot2 plot}
}
\value{
A plot with a specific theme
}
\description{
Apply a ggplot2 theme to a plot
}
\examples{
p <- mtcars |>
ggplot(aes(x = hp, y = mpg, color = cyl)) + geom_point()
my_theme(p)
}
