% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.r
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object
}
\description{
Custom ggplot2 theme
}
\examples{
iris |>
 ggplot(aes(x = Sepal.Length, y = Sepal.Width)) +
 geom_point() +
 my_theme()
}
