% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.r
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 Theme}
\usage{
my_theme()
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot2 Theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp)) +
  geom_point() +
  my_theme()
}
