% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{Experimental Function}
\usage{
myfun(x)
}
\arguments{
\item{x}{A vector of real numbers}
}
\value{
A squared vector
}
\description{
Experimental Function
}
\examples{
myfun(1:10)
}
