% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.r
\name{myfunc}
\alias{myfunc}
\title{My first function}
\usage{
myfunc(a, b)
}
\arguments{
\item{a}{A numerical vector.}

\item{b}{Also a numerical vector.}
}
\value{
A numerical vector of a + b * a.
}
\description{
My first function
}
\examples{
myfunc(3, 5)
}
