% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIBTSdata.R
\name{readIBTSData}
\alias{readIBTSData}
\title{readIBTSData}
\usage{
readIBTSData(survey = "NS-IBTS", year, quarter, species)
}
\arguments{
\item{year}{year of survey}

\item{quarter}{quarter of survey}

\item{species}{species of interest}
}
\description{
readIBTSData
}
\examples{

}
