% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeData}
\alias{removeData}
\title{removeData}
\usage{
removeData(year, quarter, species, dat, whatToRemove, lengthDivision,
  samplesWithinEachIntervall = 1)
}
\arguments{
\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{species}{The species of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
