% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeData}
\alias{removeData}
\title{removeData}
\usage{
removeData(year, quarter, species, dat, removeProcedure = 1, whatToRemove,
  doNotRemoveAbove, lengthDivision, propRemove = NULL)
}
\arguments{
\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{species}{The species of interest.}

\item{removeProcedure}{Removal procedure, given as an integer. removeProcedure==1 means that data is removed at random.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
