% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{say_hello}
\alias{say_hello}
\title{Say hello to someone}
\usage{
say_hello(name)
}
\arguments{
\item{name}{A character string of the person's name.}
}
\description{
This function prints a hello message to the specified person.
}
\examples{
say_hello("Alice")
}
