% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{simTrawlHaulsHiearchical}
\alias{simTrawlHaulsHiearchical}
\title{simTrawlHaulsHiearchical}
\usage{
simTrawlHaulsHiearchical(RFA, year, quarter, hl, ca,
  hierarchy = c("StatRec", "haul.id"), selection = c("S", "R"))
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{hl}{data formated as DATRAS HL table}

\item{ca}{data formated as DATRAS CA table}
}
\description{
Simulates selection of hauls by selection of StatRec without replacement, and selection of haul.id with replacement
}
