% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{square}
\alias{square}
\title{Add together two numbers.}
\usage{
square(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The square of \code{x}.
}
\description{
Add together two numbers.
}
\examples{
#' square(2)
}
