% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{split a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{a character vector}

\item{split}{split character}
}
\value{
A char vectir
}
\description{
split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")

}
