% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage-package}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: Demonstration of Package Creation}
\description{
Not a whole lot, just learning!
}
\author{
\strong{Maintainer}: Adim Okwudishu \email{adimkris@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
