% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage-package}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: Demos creating a package}
\description{
What the package does (not very much right now).
}
\author{
\strong{Maintainer}: Chidi Agbaeruneke \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
