% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage-package}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: Demo creating a package for 413 Data Science}
\description{
Learning how to build a package by creating a test package.
}
\author{
\strong{Maintainer}: Christopher Hoffmann \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
