% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackage-package.R
\docType{package}
\name{testPackage-package}
\alias{testPackage}
\alias{testPackage-package}
\title{testPackage: Test Package, First One Yay!}
\description{
I don't know what this does. :)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fai-k/testPackage}
  \item Report bugs at \url{https://github.com/fai-k/testPackage/issues}
}

}
\author{
\strong{Maintainer}: T K \email{email.me@gmail.com} (\href{https://orcid.org/1111-2222-3333-4444}{ORCID})

}
\keyword{internal}
