% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPackage-package.R
\docType{package}
\name{TestPackage-package}
\alias{TestPackage}
\alias{TestPackage-package}
\title{TestPackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
