% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage-package}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: My Package as a Demo}
\description{
This is my basic package.
}
\author{
\strong{Maintainer}: Olivia Freides \email{oliviafreides@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
