% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testpackage-package.R
\docType{package}
\name{Testpackage-package}
\alias{Testpackage}
\alias{Testpackage-package}
\title{Testpackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Poolpoul/Testpackage}
  \item \url{https://poolpoul.github.io/Testpackage/}
  \item Report bugs at \url{https://github.com/Poolpoul/Testpackage/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
