\name{testPackage-package}
\alias{testPackage-package}
\alias{testPackage}
\docType{package}
\title{
\packageTitle{testPackage}
}
\description{
\packageDescription{testPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testPackage}
\packageIndices{testPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testPackage}

Maintainer: \packageMaintainer{testPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
