\name{TestPackage-package}
\alias{TestPackage-package}
\alias{TestPackage}
\docType{package}
\title{
\packageTitle{TestPackage}
}
\description{
\packageDescription{TestPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestPackage}
\packageIndices{TestPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TestPackage}

Maintainer: \packageMaintainer{TestPackage}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
