% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage.R
\name{testpackage}
\alias{testpackage}
\title{A Test Function}
\usage{
testpackage(param_test = TRUE)
}
\arguments{
\item{love}{Do you want a test? Defaults to TRUE.}
}
\description{
This function allows you to express a test.
}
\examples{
testpackage()
}
\keyword{param_test}
