% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage.R
\docType{package}
\name{testpackage}
\alias{testpackage}
\alias{testpackage-package}
\title{\code{testpackage} package}
\description{
This is a testpackage for trying to debug the travis build of GrasslandAllocatr
}
