% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage.R
\docType{package}
\name{testpackage}
\alias{testpackage}
\title{testpackage: Learnin' about UFOs}
\description{
Data about UFO sightings in the United States. Functions for plotting by
state and importing the data.
}
\section{Data import functions}{

\code{\link{ufo_data}}
}

\section{Plotting functions}{

\code{\link{plotbymonth}}
}

\section{Other section can go here, e.g. thoughts on future dev}{

\itemize{
  \item first item
  \item second item
}
}

