% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_HSRC.R
\name{theme_HSRC}
\alias{theme_HSRC}
\title{Create an HSRC ggplot theme}
\usage{
theme_HSRC(
  base_size = 18,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170
)
}
\arguments{
\item{base_size}{Base size for text elements. Default is 18}

\item{base_family}{Base font family}

\item{base_line_size}{Base line size for line elements. Default is base_size / 170}

\item{base_rect_size}{Base size for rect elements. Default is base_size / 170}
}
\value{
A plot with text (left) and logo (right)
}
\description{
\code{theme_HSRC} is a theme based on theme_minimal. No grid lines, only x and y axis.
Horizontal y-axis label
Dark grey rather than black for text elements
}
\examples{
library(ggplot2)
pl_1 <- ggplot(mpg, aes(displ, hwy, colour = class)) + geom_point()
pl_1_no_legend <- ggplot(mpg, aes(displ, hwy, colour = class)) +
geom_point(show.legend = FALSE)

pl_1
pl_1 + theme_HSRC()
pl_1_no_legend
pl_1_no_legend + theme_HSRC()

}
